### 2020-01-20 #359 @llrs
Submission: BaseSet 
Submitting Author: Lluís (@llrs)
Repository:   [llrs/BaseSet](https://github.com/llrs/BaseSet)
Version submitted:   0.0.10
Version accepted: TBD

---

```
Package: BaseSet
Title: Provides classes for working with sets
Version: 0.0.10
    person(given = "Lluís ",
           family = "Revilla Sancho",
           email = "lluis.revilla@gmail.com")
Description: A set collection, while not "tidy" in itself, can
    be thought of as three tidy data frames describing sets, elements and
    relations respectively. 'BaseSet' provides an approach to manipulate,
    load and use these virtual data frames.
    xml2
    BiocStyle,
    covr,
    forcats,
    GO.db,
    GSEABase,
    org.Hs.eg.db,
    reactome.db,
    Biobase
    knitr
Collate:
    'AllClasses.R'
    'AllGenerics.R'
    'GMT.R'
    'GeneSetCollection.R'
    'activate.R'
    'add.R'
    'add_column.R'
    'add_relation.R'
    'adjacency.R'
    'arrange.R'
    'basesets-package.R'
    'cartesian.R'
    'complement.R'
    'data_frame.R'
    'deactivate.R'
    'droplevels.R'
    'elements.R'
    'filter.R'
    'group.R'
    'group_by.R'
    'head.R'
    'incidence.R'
    'independent.R'
    'operations.R'
    'intersection.R'
    'length.R'
    'list.R'
    'move_to.R'
    'mutate.R'
    'names.R'
    'naming.R'
    'nested.R'
    'obo.R'
    'power_set.R'
    'print.R'
    'pull.R'
    'relations.R'
    'remove.R'
    'remove_column.R'
    'rename.R'
    'select.R'
    'set.R'
    'size.R'
    'subtract.R'
    'tidy-set.R'
    'union.R'
    'utils-pipe.R'
    'xml.R'
    'zzz.R'

```

	- [ ] workflow automataion

The package implements methods to work on sets, doing intersection, union, complementary and other set operations in a "tidy" way. It also allows to import from several formats used in the life science world. Like the GMT and the GAF or the OBO format file for ontologies.

The idea is to use the package for working with sets and signatures of genes in scRNAseq or in pathways and ontologies but it might work with other fields.

There is the [sets]( https://CRAN.R-project.org/package=sets) package which implements a more generalized approach, that can store functions or lists as an element of a set (while mine it only allows to store a character or factor), but it is harder to operate in a tidy/long way. Also the operations of intersection and union need to happen between two different objects, while TidySet objects (the class implemented in BaseSet) can store a single set or thousands of them.
In BaseSet is easier to operate and implement new fuzzy logic operations. It is developed openly on github compared to sets which I couldn't track how it is being developed.

The [GSEABase](https://bioconductor.org/packages/GSEABase) partially implements this, but it doesn't allow to store fuzzy sets and it is also quite slow as it creates several classes for annotating each set. Neither does the [BiocSets](https://bioconductor.org/packages/BiocSet) the package, which don't use the fuzzy set logic.

There is also the [hierarchicalSets](https://CRAN.R-project.org/package=hierarchicalSets) package that is focused on clustering of sets that are inside other sets and visualizations. However, BaseSet is focused on storing and manipulate sets including hierarchical sets.

Most of the replies are copied from #339, handeled by @melvidoni.
