### 2020-01-28 #359 @annakrystalli
Hello again @llrs!

Thanks for your patience with the editors checks. I've been traveling but I've also been having some issues with the checks which I'm having to seek some help with.  Will be with you shortly, if not with the full checks, at least with those that I am able to perform successfully.
