### 2020-01-29 #359 @annakrystalli
### Editor checks:

---

### Editor comments

#### Installation issues

When I try to do a full install and build vignettes, some suggested bioconductor dependencies are not installing and causing errors when attempting to build the vignettes.

``` r
devtools::install_github("llrs/BaseSet", dependencies = T, build_vignettes = T)
```

```
#> Installing 7 packages: BiocStyle, GO.db, GSEABase, org.Hs.eg.db, reactome.db, Biobase, BH
#> Error: (converted from warning) packages 'BiocStyle', 'GO.db', 'GSEABase', 'org.Hs.eg.db', 'reactome.db', 'Biobase' are not available (for R version 3.6.0)

***

   Error: Vignette re-building failed.
   Execution halted
Error in (function (command = NULL, args = character(), error_on_status = TRUE,  :
  System command error

```
Reviewevers as well as potential future collaborators that will need to run `devtools::check()` and `devtools::document()`, need a section in the `README` or a separate `CONTRIBUTING.md` documenting the installation instructions required to install all dependencies (including suggested).

eg:

```r
BiocManager::install("BiocStyle")
BiocManager::install("org.Hs.eg.db", type = "source")
BiocManager::install("GO.db", type = "source")
BiocManager::install("reactome.db", type = "source")

# test that package install and docs build successfully
devtools::install_github("llrs/BaseSet", dependencies = T, build_vignettes = T, force = T)
```

You can check out more of the adventures I had troubleshooting in [this issue comment](https://github.com/ropensci/dev_guide/issues/210#issuecomment-579535101) for more details. If you have some extra insight (as I imagine you are more familiar with bioconductor) to add to the discussion it would be more than welcome!

#### `check()` and `test()`

All good

#### `goodpractice::gp()` output

A few flags for minor formatting.
```
It is good practice to

  ✖ avoid long code lines, it is bad for
    readability. Also, many people prefer editor windows
    that are about 80 characters wide. Try make your
    lines shorter than 80 characters

    R/AllClasses.R:105:1
    R/AllGenerics.R:149:1
    R/AllGenerics.R:163:1
    R/length.R:171:1
    R/obo.R:122:1
    ... and 16 more lines
```

This is actually a bit more important. Is there a reason you are importing `methods` and `utils` packages as a whole?

```
  ✖ not import packages as a whole, as this
    can cause name clashes between the imported
    packages. Instead, import only the specific
    functions you need.
────────────────────────────────────────────────────────────────
```

#### Test coverage

Test coverage is overall good but is there a reason there are some files that are not tested at all or below 75%?

See the files indicated below:

```
BaseSet Coverage: 89.13%
R/xml.R: 0.00%
R/elements.R: 62.50%
R/set.R: 62.50%
R/size.R: 71.43%
R/pull.R: 73.33%

```

@llrs, could you please add the rOpenSci under review badge to your README?

Other than these minor issues, I am happy to start seeking reviewers.

---

Reviewers:
Due date:
