### 2020-01-29 #359 @llrs
Many thanks for your review @annakrystalli

- Installation problems: Oh, sorry on travis and my computer I didn't had these installation problems. I would follow the BiocViews solution, but usually the easier solution might be to use `BiocManager::install("llrs/BaseSet")`. Lately there has been some problems with the Bioconductor binaries for Windows related to BiocStyle (Some of your problems might come from this, but I'm not sure as I have not submitted a data package).
- Regarding long lines: Some of I had considered them easier to read with slightly over 80 character per line, some of them will be me forgetting to reformat them. I'll check each one of those lines.
- `{methods}` & `{utils}`: I thought that it would be easier to work with the S4 class if I imported both packages and I use many functions of them. It is easier for me while writing the package too. They are both installed by default so I did not expect much problems, but I will import only the functions needed.
- Test coverage: I forgot to delete the xml.R file. I think there isn't a compromise on which format are written with sets on `.xml` files. The other few packages that are below 75% might be that I didn't check for some exceptions.
- rOpenSci badge: Will add it.

I'll modify the package to address your points by the beginning of next week I'll have them addressed. I'll let you know when I updated the package.
