### 2020-01-29 #359 @annakrystalli
Hello again @llrs,

Aha, it seems I have been trying to install the package the wrong way all this time, by using `devtools::use_github()` as I normally do for editors checks when I should be using `BiocManager::install()`

However, when I try to install using `BiocManager::install("llrs/BaseSet", dependencies = TRUE, build_vignettes = TRUE)` I still get the same errors for a couple of suggested bioconductor packages I removed to test it out.

I'm really intrigued by the fact that the packages builds successfully for you on TRAVIS. Looking at your `travis.yml` I see you are testing on `r: bioc-devel` only. I think this where the testing computational environment and my own (and of reviewers and future collaborators) would differ.  I feel also testing on `r: release` at the very least would give a better idea of what challenges others would encounter. Would you mind trying that out for me?
