### 2020-01-30 #359 @annakrystalli
So the errors I am getting are the same as I was previously with using `devtools::install_github()`, namely, any missing bioconductor packages that need installing from source are missing. I've deleted package `GO.db` from my system and so now,  the vignette that loads it is failing. Here's a condensed error output:

```r
BiocManager::install("llrs/BaseSet", dependencies = TRUE, build_vignettes = TRUE)
```

```r
E  creating vignettes (8.6s)
   --- re-building ‘advanced.Rmd’ using rmarkdown

   Attaching package: 'dplyr'

.
.
.

   Quitting from lines 49-56 (advanced.Rmd)
   Error: processing vignette 'advanced.Rmd' failed with diagnostics:
   there is no package called 'GO.db'
   --- failed re-building ‘advanced.Rmd’

   --- re-building ‘basic.Rmd’ using rmarkdown
   --- finished re-building ‘basic.Rmd’

   --- re-building ‘fuzzy.Rmd’ using rmarkdown
   --- finished re-building ‘fuzzy.Rmd’

   SUMMARY: processing the following file failed:
     ‘advanced.Rmd’

   Error: Vignette re-building failed.
   Execution halted
Error: Failed to install 'BaseSet' from GitHub:
  System command error, exit status: 1, stdout + stderr (last 10 lines):
E> --- finished re-building ‘basic.Rmd’
E>
E> --- re-building ‘fuzzy.Rmd’ using rmarkdown
E> --- finished re-building ‘fuzzy.Rmd’
E>
E> SUMMARY: processing the following file failed:
E>   ‘advanced.Rmd’
E>
E> Error: Vignette re-building failed.
E> Execution halted
```

The reason this would not fail on your system is because you already have all the necessary package installed, but reviewers or new contributors won't necessarily. The reason I think the travis tests are passing is because of you are testing on `bioc-devel`. I can't think of why else installation on TRAVIS works while it doesn't for me. Testing on `r: release` feels like it would give you a more realistic situtation of what manual install steps would be required at any point by developers/ reviewers to be able to fully test and check the package.

Out of interest, you could try removing `GO.db` locally on your system and see what happens. For info, here's my session info too, ie I'm on a mac not windows.

```r
> session_info()
─ Session info ───────────────────────────────────────────────────────────────────────
 setting  value
 version  R version 3.6.0 (2019-04-26)
 os       macOS Mojave 10.14.3
 system   x86_64, darwin15.6.0
 ui       RStudio
 language (EN)
 collate  en_GB.UTF-8
 ctype    en_GB.UTF-8
 tz       America/Los_Angeles
 date     2020-01-30

─ Packages ───────────────────────────────────────────────────────────────────────────
 package     * version     date       lib source
 assertthat    0.2.1       2019-03-21 [1] CRAN (R 3.6.0)
 backports     1.1.5       2019-10-02 [1] CRAN (R 3.6.0)
 BiocManager   1.30.10     2019-11-16 [1] CRAN (R 3.6.0)
 callr         3.4.1       2020-01-24 [1] CRAN (R 3.6.0)
 cli           2.0.1       2020-01-08 [1] CRAN (R 3.6.0)
 clipr         0.7.0       2019-07-23 [1] CRAN (R 3.6.0)
 crayon        1.3.4       2017-09-16 [1] CRAN (R 3.6.0)
 curl          4.3         2019-12-02 [1] CRAN (R 3.6.0)
 desc          1.2.0       2018-05-01 [1] CRAN (R 3.6.0)
 devtools    * 2.2.1       2019-09-24 [1] CRAN (R 3.6.0)
 digest        0.6.23      2019-11-23 [1] CRAN (R 3.6.0)
 ellipsis      0.3.0       2019-09-20 [1] CRAN (R 3.6.0)
 evaluate      0.14        2019-05-28 [1] CRAN (R 3.6.0)
 fansi         0.4.1       2020-01-08 [1] CRAN (R 3.6.0)
 fs            1.3.1       2019-05-06 [1] CRAN (R 3.6.0)
 glue          1.3.1       2019-03-12 [1] CRAN (R 3.6.0)
 htmltools     0.4.0       2019-10-04 [1] CRAN (R 3.6.0)
 knitr         1.27        2020-01-16 [1] CRAN (R 3.6.0)
 magrittr      1.5         2014-11-22 [1] CRAN (R 3.6.0)
 memoise       1.1.0       2017-04-21 [1] CRAN (R 3.6.0)
 packrat       0.5.0       2018-11-14 [1] CRAN (R 3.6.0)
 pkgbuild      1.0.6       2019-10-09 [1] CRAN (R 3.6.0)
 pkgload       1.0.2       2018-10-29 [1] CRAN (R 3.6.0)
 prettyunits   1.1.1       2020-01-24 [1] CRAN (R 3.6.0)
 processx      3.4.1       2019-07-18 [1] CRAN (R 3.6.0)
 ps            1.3.0       2018-12-21 [1] CRAN (R 3.6.0)
 R6            2.4.1       2019-11-12 [1] CRAN (R 3.6.0)
 Rcpp          1.0.3       2019-11-08 [1] CRAN (R 3.6.0)
 remotes       2.1.0       2019-06-24 [1] CRAN (R 3.6.0)
 reprex      * 0.3.0       2019-05-16 [1] CRAN (R 3.6.0)
 rlang         0.4.3       2020-01-24 [1] CRAN (R 3.6.0)
 rmarkdown     2.1         2020-01-20 [1] CRAN (R 3.6.0)
 rprojroot     1.3-2       2018-01-03 [1] CRAN (R 3.6.0)
 rstudioapi    0.10.0-9000 2019-05-30 [1] Github (rstudio/rstudioapi@31d1afa)
 sessioninfo   1.1.1       2018-11-05 [1] CRAN (R 3.6.0)
 testthat    * 2.3.1       2019-12-01 [1] CRAN (R 3.6.0)
 usethis     * 1.5.1.9000  2020-01-29 [1] Github (r-lib/usethis@4194fd6)
 whisker       0.4         2019-08-28 [1] CRAN (R 3.6.0)
 withr         2.1.2       2018-03-15 [1] CRAN (R 3.6.0)
 xfun          0.12        2020-01-13 [1] CRAN (R 3.6.0)

[1] /Library/Frameworks/R.framework/Versions/3.6/Resources/library
```
