+++ R/basesets-package.R
@@ -1,3 +1,3 @@
 #' @details
-#' It provides a class [`TidySet`]s with lots of methods to do operations with sets
+#' It provides a class [`TidySet`] with methods to do operations with sets.
 "_PACKAGE"
