+++ R/mutate.R
@@ -21,7 +21,8 @@ dplyr::mutate
 #' a <- tidySet(relations)
 #' a <- mutate_element(a, Type = c(rep("Gene", 4), rep("lncRNA", 2)))
 #' a
-#' b <- mutate_relation(a, Type = sample(c("PPI", "PF", "MP"), 12,  replace = TRUE))
+#' b <- mutate_relation(a, Type = sample(c("PPI", "PF", "MP"), 12,
+#'                                          replace = TRUE))
 #' @rdname mutate_
 #' @export
 #' @method mutate TidySet
