+++ R/naming.R
@@ -28,8 +28,8 @@ check_symbol <- function(symbol) {
 #' @param start,middle Character used as a start symbol or to divide
 #' \code{sets1} and \code{sets2}.
 #' @param sets1,sets2 Character of sets
-#' @param collapse_symbol Name of the symbol that joins the sets on \code{sets1} and
-#' \code{sets2}.
+#' @param collapse_symbol Name of the symbol that joins the sets on
+#' \code{sets1} and \code{sets2}.
 #' @seealso \code{\link{set_symbols}}
 #' @return A character vector combining the sets
 #' @export
