+++ README.Rmd
@@ -117,5 +117,5 @@ fuzzy_set %>%
 ```
 
 
-Please note that the 'BaseSet' project is released with a  [Contributor Code of Conduct](CODE_OF_CONDUCT.md).
+Please note that the BaseSet project is released with a  [Contributor Code of Conduct](CODE_OF_CONDUCT.md).
 By contributing to this project, you agree to abide by its terms.
