### 2020-01-31 6697a45
Increase test coverage
