+++ tests/testthat/test-set_size.R
@@ -43,3 +43,18 @@ test_that("set_size works well", {
   expect_equal(sum(bs$probability), 1L)
   expect_equal(bs, bs2)
 })
+
+
+test_that("set_size works well", {
+  signature <- list(Tcell = c("CD4" = 1, "CD8" = 0.5,
+                              "CD45" = 0.75, "Tbet" = 0.2),
+                    Bcell = c("CD45" = 1, "CD19" = 0.5,
+                              "CD40" = 1, "IgM" = 0.8))
+
+  cells <- tidySet(signature)
+  b <- activate(cells, "sets")
+  bs2 <- size(b, "Bcell")
+  expect_equal(sum(bs2$probability), 1L)
+})
+
+
