+++ R/AllGenerics.R
@@ -1,5 +1,11 @@
 #' @include AllClasses.R
-#' @import methods
+#' @importFrom methods setGeneric
+#' @importFrom methods as
+#' @importFrom methods new
+#' @importFrom methods slot
+#' @importFrom methods slot<-
+#' @importFrom methods validObject
+#' @importFrom methods show
 NULL
 
 #' Elements of the TidySet
@@ -16,7 +22,7 @@ NULL
 #' @examples
 #' TS <- tidySet(list(A = letters[1:5], B = letters[2:10]))
 #' elements(TS)
-setGeneric("elements", function(object) standardGeneric("elements"))
+methods::setGeneric("elements", function(object) setGeneric("elements"))
 
 #' Sets of the TidySet
 #'
@@ -32,7 +38,7 @@ setGeneric("elements", function(object) standardGeneric("elements"))
 #' @examples
 #' TS <- tidySet(list(A = letters[1:5], B = letters[2:10]))
 #' sets(TS)
-setGeneric("sets", function(object) standardGeneric("sets"))
+methods::setGeneric("sets", function(object) standardGeneric("sets"))
 
 #' Check if a TidySet is fuzzy.
 #'
@@ -42,7 +48,7 @@ setGeneric("sets", function(object) standardGeneric("sets"))
 #' @return A logical value.
 #' @export is.fuzzy
 #' @family methods
-setGeneric("is.fuzzy", function(object) standardGeneric("is.fuzzy"))
+methods::setGeneric("is.fuzzy", function(object) standardGeneric("is.fuzzy"))
 
 #' Number of sets
 #'
@@ -55,7 +61,7 @@ setGeneric("is.fuzzy", function(object) standardGeneric("is.fuzzy"))
 #' @examples
 #' TS <- tidySet(list(A = letters[1:2], B = letters[5:7]))
 #' nSets(TS)
-setGeneric("nSets", function(object) standardGeneric("nSets"))
+methods::setGeneric("nSets", function(object) standardGeneric("nSets"))
 
 #' Number of elements
 #'
@@ -68,7 +74,7 @@ setGeneric("nSets", function(object) standardGeneric("nSets"))
 #' @examples
 #' TS <- tidySet(list(A = letters[1:2], B = letters[5:7]))
 #' nElements(TS)
-setGeneric("nElements", function(object) standardGeneric("nElements"))
+methods::setGeneric("nElements", function(object) standardGeneric("nElements"))
 
 #' Number of relations
 #'
@@ -81,7 +87,7 @@ setGeneric("nElements", function(object) standardGeneric("nElements"))
 #' @examples
 #' TS <- tidySet(list(A = letters[1:2], B = letters[5:7]))
 #' nRelations(TS)
-setGeneric("nRelations", function(object) standardGeneric("nRelations"))
+methods::setGeneric("nRelations", function(object) standardGeneric("nRelations"))
 
 #' Incidence
 #'
@@ -90,7 +96,7 @@ setGeneric("nRelations", function(object) standardGeneric("nRelations"))
 #' @export incidence
 #' @family methods
 #' @seealso \code{\link{adjacency}}
-setGeneric("incidence", function(object) standardGeneric("incidence"))
+methods::setGeneric("incidence", function(object) standardGeneric("incidence"))
 
 #' Relations of the TidySet
 #'
@@ -107,21 +113,21 @@ setGeneric("incidence", function(object) standardGeneric("incidence"))
 #' @examples
 #' TS <- tidySet(list(A = letters[1:2], B = letters[5:7]))
 #' relations(TS)
-setGeneric("relations", function(object) standardGeneric("relations"))
+methods::setGeneric("relations", function(object) standardGeneric("relations"))
 
[…]
