+++ NAMESPACE
@@ -159,8 +159,6 @@ exportMethods(set_size)
 exportMethods(sets)
 exportMethods(show)
 exportMethods(subtract)
-import(methods)
-import(utils)
 importFrom(dplyr,arrange)
 importFrom(dplyr,filter)
 importFrom(dplyr,group_by)
@@ -169,7 +167,14 @@ importFrom(dplyr,n_distinct)
 importFrom(dplyr,pull)
 importFrom(dplyr,select)
 importFrom(magrittr,"%>%")
+importFrom(methods,"slot<-")
+importFrom(methods,as)
+importFrom(methods,new)
 importFrom(methods,setClassUnion)
+importFrom(methods,setGeneric)
+importFrom(methods,show)
+importFrom(methods,slot)
+importFrom(methods,validObject)
 importFrom(rlang,"!!")
 importFrom(rlang,enquo)
 importFrom(rlang,exprs)
