### 2020-02-04 ea4dd59
add r release build
This build will test whether manual instructions are sufficient to install all dependencies locally (for reviewers and collaborators)
