+++ .travis.yml
@@ -4,10 +4,14 @@ language: R
 sudo: false
 cache: packages
 
+jobs:
+  include:
+    - r: bioc-release
+    - r: release
+      script: Rscript -e `remotes::install_cran("BiocManager"); BiocManager::install(c("Biobase", "GO.db", "GSEABase", "org.HS.eg.db", 
+                         "reactome.db", "BiocStyle"), type = "source")`
 after_success:
     - Rscript -e 'covr::codecov()'
-
-r: bioc-release
 
 notifications: # To avoid filling the inbox
   email:
