### 2020-02-04 #359 @annakrystalli
Hi @llrs,

Sorry, I probably didn't make myself clear enough. For sure you should be testing on `bioc-release` (perhaps on `bioc-devel` also). What I was suggesting is that you run a matrix of jobs that include testing on `r-release` and including any manual setup code during the build to both test that any awkward dependencies required for a reviewer/collaborator to work with the source code locally have:
a) been identified
b) their installation has been successfully specified.

Does that make sense? I've opened a PR to your repo with my attempt at implementing this. Let's see if it's successful!
