### 2020-02-04 #359 @llrs
Well, my impression was that CI should mimic the steps done on CRAN, not for the users. The steps are different, given that CRAN installs Bioconductor packages to be able to build packages which depend on Bioconductor.
That's why I modified the .travis.yml to only test on bioc-release, which is also what users need in order to install BaseSet. But it makes sense also to test the specific steps that users need to do.

I will try to modify the github actions to test also on OS, windows and Linux in both conditions to be sure.
