### 2020-02-07 f2fb91f
Update building on github actions
