### 2020-02-10 #359 @annakrystalli
Hello again @llrs 👋

I'm still looking for reviewers currently. I had a potential reviewer respond recently that sadly they did not currently have time at the moment but they also made the following comment:
> the description in the Readme doesn't give me enough information to figure out what this package is about.

I generally agree with the comment and was imagining it would be picked up during review but I now think it's probably best to deal with it now as it might be making it difficult to find reviewers.

So could you add a little more detail to the README as well as the description in the `DESCRIPTION` file?  In particular, the information you have included in here in the package submission gives a good idea of the strengths and domains of the package which are not so clear currently in the `README`.
