### 2020-02-10 79ed3b9
Add new section comparing with other packages and expanding the description of the packages
