+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: BaseSet
 Title: Working with sets the tidy way
-Version: 0.0.10
+Version: 0.0.11
 Authors@R: 
     person(given = "Lluís ",
            family = "Revilla Sancho",
