### 2020-02-10 #359 @llrs
Hi @annakrystalli,

Thanks for sending the feedback, I appreciate it. I've edited both the DESCRIPTION and the README. Hope this way it will be easier to get reviewers.
