### 2020-02-10 a18fdbb
Use BiocManager instead of remotes
