### 2020-02-11 649de65
Add sets6 package and add more links on the README
