+++ README.md
@@ -27,6 +27,282 @@ It also allows to import from several formats used in the life science
 world. Like the GMT and the GAF or the [OBO
 format](http://www.obofoundry.org/) file for ontologies.
 
+You can save information about the elements, sets and their relationship
+on the object itself. For instance origin of the set, categorical or
+numeric data associated with sets…
+
+Watch BaseSet working on the [examples](#Examples) below and in the
+vignettes. You can also find [related packages](#related) and the
+differences with BaseSet. If you have some questions or bugs [open an
+issue](https://github.com/llrs/BaseSet/issues) (remember of the [Code of
+Conduct](#CoC))
+
+# Installation
+
+Before installation you might need to install some of the suggested
+packages from Bioconductor:
+
+``` r
+if (!require("BiocManager")) {
+  install.packages("BiocManager")
+  BiocManager::install(c("Biobase", "GO.db", "GSEABase", "org.HS.eg.db", 
+                         "reactome.db", "BiocStyle"), type = "source")
+}
+```
+
+You can install the latest version of BaseSet from
+[Github](https://github.com/llrs/BaseSet) with:
+
+``` r
+BiocManager::install("llrs/BaseSet", 
+                     dependencies = TRUE, build_vignettes = TRUE, force = TRUE)
+```
+
+# Examples
+
+## Sets
+
+We can create a set like this:
+
+``` r
+sets <- list(A = letters[1:5], B = c("a", "f"))
+sets_analysis <- tidySet(sets)
+sets_analysis
+#>   elements sets fuzzy
+#> 1        a    A     1
+#> 2        a    B     1
+#> 3        b    A     1
+#> 4        c    A     1
+#> 5        d    A     1
+#> 6        e    A     1
+#> 7        f    B     1
+```
+
+Perform typical operations like union, intersection. You can name the
+resulting set or let the default name:
+
+``` r
+union(sets_analysis, sets = c("A", "B")) 
+#>   elements sets fuzzy
+#> 1        a  A∪B     1
+#> 2        b  A∪B     1
+#> 3        c  A∪B     1
+#> 4        d  A∪B     1
+#> 5        e  A∪B     1
+#> 6        f  A∪B     1
+# Or we can give a name to the new set
+union(sets_analysis, sets = c("A", "B"), name = "D")
+#>   elements sets fuzzy
+#> 1        a    D     1
+#> 2        b    D     1
+#> 3        c    D     1
+#> 4        d    D     1
+#> 5        e    D     1
+#> 6        f    D     1
+# Or the intersection
+intersection(sets_analysis, sets = c("A", "B"))
+#>   elements sets fuzzy
+#> 1        a  A∩B     1
+# Keeping the other sets:
+intersection(sets_analysis, sets = c("A", "B"), name = "D", keep = TRUE) 
+#>   elements sets fuzzy
+#> 1        a    A     1
+#> 2        a    B     1
+#> 3        a    D     1
+#> 4        b    A     1
+#> 5        c    A     1
+#> 6        d    A     1
+#> 7        e    A     1
+#> 8        f    B     1
+```
+
+And compute size of sets among other things:
+
+``` r
+set_size(sets_analysis)
+#>   sets size probability
+#> 1    A    5           1
+#> 2    B    2           1
+```
+
+The elements in one set not present in other:
+
+``` r
[…]
