+++ README.Rmd
@@ -30,23 +30,11 @@ The goal of BaseSet is to facilitate working with sets in an efficient way. The 
 
 It also allows to import from several formats used in the life science world. Like the GMT and the GAF or the [OBO format](http://www.obofoundry.org/) file for ontologies.
 
-# Related packages
+You can save information about the elements, sets and their relationship on the object itself. For instance origin of the set, categorical or numeric data associated with sets...
 
-There are several other packages related to sets, which partially overlap with BaseSet functionality:
+Watch BaseSet working on the [examples](#Examples) below and in the vignettes. You can also find [related packages](#related) and the differences with BaseSet. If you have some questions or bugs [open an issue](https://github.com/llrs/BaseSet/issues) (remember of the [Code of Conduct](#CoC))
 
- - [sets]( https://CRAN.R-project.org/package=sets)  
- Implements a more generalized approach, that can store functions or lists as an element of a set (while BaseSet only allows to store a character or factor), but it is harder to operate in a tidy/long way. Also the operations of intersection and union need to happen between two different objects, while a single TidySet object (the class implemented in BaseSet) can store one or thousands of sets.
-
- - [GSEABase](https://bioconductor.org/packages/GSEABase)  
- Implements a class to store sets and related information, but it doesn't allow to store fuzzy sets and it is also quite slow as it creates several classes for annotating each set. 
-  
- - [BiocSets](https://bioconductor.org/packages/BiocSets)  
- Implements a tidy class for sets but does not handle fuzzy sets. It also has less functionality to operate with sets, like power sets and cartesian product. BiocSets was influenced by the development of this package. 
-
- - [hierarchicalSets](https://CRAN.R-project.org/package=hierarchicalSets)  
- This package is focused on clustering of sets that are inside other sets and visualizations. However, BaseSet is focused on storing and manipulate sets including hierarchical sets.
-
-## Installation
+# Installation
 
 Before installation you might need to install some of the suggested packages from Bioconductor:
 
@@ -66,13 +54,15 @@ BiocManager::install("llrs/BaseSet",
 ```
 
  
-## Examples
+# Examples {#Examples}
 
 ```{r include=FALSE}
 library("BaseSet")
 ```
 
-To work with sets we need to have the relations between the elements and the sets:
+## Sets
+
+We can create a set like this:
 
 ```{r TidySet}
 sets <- list(A = letters[1:5], B = c("a", "f"))
@@ -80,16 +70,16 @@ sets_analysis <- tidySet(sets)
 sets_analysis
 ```
 
-Perform typical operations like union, intersection:
+Perform typical operations like union, intersection. You can name the resulting set or let the default name:
 
 ```{r union-intersection}
 union(sets_analysis, sets = c("A", "B")) 
-# Or we can give a name to the new set and keep the others sets
+# Or we can give a name to the new set
 union(sets_analysis, sets = c("A", "B"), name = "D")
-# We can do the same in the intersection
-intersection(sets_analysis, sets = c("A", "B"), name = "D") 
-# Or we can omit the new name: 
+# Or the intersection
 intersection(sets_analysis, sets = c("A", "B"))
+# Keeping the other sets:
+intersection(sets_analysis, sets = c("A", "B"), name = "D", keep = TRUE) 
 ```
 
 And compute size of sets among other things:
@@ -104,21 +94,28 @@ The elements in one set not present in other:
[…]
