### 2020-02-13 c02174c
Reduce installation of dependencies
It tried to download packages from github
