### 2020-02-13 a06100b
Simplify yaml
