+++ .github/workflows/R-CMD-check.yaml
@@ -3,30 +3,14 @@ on: [push, pull_request]
 name: R-CMD-check
 
 jobs:
-  macOS-check:
-    runs-on: macOS-latest
+  R-CMD-check:
+    runs-on: [macOS-latest, ubuntu-latest, windows-latest]
     steps:
       - uses: actions/checkout@v1
       - uses: r-lib/actions/setup-r@master
-      - name: Install dependencies
-        run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'BiocManager'))" -e 'BiocManager::install("llrs/BaseSet", dependencies = TRUE)'
-      - name: Check
-        run: Rscript -e "rcmdcheck::rcmdcheck(args = '--no-manual', error_on = 'error')"
-  windows-check:
-    runs-on: windows-latest
-    steps:
-      - uses: actions/checkout@v1
-      - uses: r-lib/actions/setup-r@master
-      - name: Install dependencies
-        run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'BiocManager'))" -e 'BiocManager::install("llrs/BaseSet", dependencies = TRUE)'
-      - name: Check
-        run: Rscript -e "rcmdcheck::rcmdcheck(args = '--no-manual', error_on = 'error')"
-  Linux-check:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v1
-      - uses: r-lib/actions/setup-r@master
-      - name: Install dependencies
-        run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'BiocManager'))" -e 'BiocManager::install("llrs/BaseSet", dependencies = TRUE)'
+      - name: Install basic dependencies
+        run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'BiocManager'))"
+      - name: Install all dependencies
+        run: Rscript -e 'BiocManager::install("llrs/BaseSet", dependencies = TRUE)'
       - name: Check
         run: Rscript -e "rcmdcheck::rcmdcheck(args = '--no-manual', error_on = 'error')"
