### 2020-02-13 29dcfa7
Use a matrix to set up the machines
