+++ .github/workflows/R-CMD-check.yaml
@@ -14,6 +14,6 @@ jobs:
       - name: Install basic dependencies
         run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'BiocManager'))"
       - name: Install all dependencies
-        run: Rscript -e 'BiocManager::install("llrs/BaseSet", dependencies = TRUE)'
+        run: Rscript -e "BiocManager::install('llrs/BaseSet', dependencies = TRUE)"
       - name: Check
         run: Rscript -e "rcmdcheck::rcmdcheck(args = '--no-manual', error_on = 'error')"
