### 2020-02-13 3d58daa
"Remove" curl from ubuntu; avoid cancelling if one task fails
