+++ .github/workflows/R-CMD-check.yaml
@@ -6,13 +6,9 @@ jobs:
   check:
     runs-on: ${{ matrix.os }}
     strategy:
+      fail-fast: false # Avoid canceling other builds if one fails.
       matrix:
         os: [macos-latest, windows-latest, ubuntu-latest]
-      include:
-        os: ubuntu-latest
-        - steps:
-          - name: curl
-            run: apt install libcurl4-openssl-dev
     steps:
       - uses: actions/checkout@v1
       - uses: r-lib/actions/setup-r@master
