### 2020-02-13 b41dd17
Update R-release configuration
