+++ .travis.yml
@@ -8,8 +8,7 @@ jobs:
   include:
     - r: bioc-release
     - r: release
-      script: Rscript -e 'remotes::install_cran("BiocManager"); BiocManager::install(c("Biobase", "GO.db", "GSEABase", "org.Hs.eg.db", 
-                         "reactome.db", "BiocStyle"), type = "source")'
+      script: Rscript -e 'remotes::install_cran("BiocManager"); BiocManager::install(c("llrs/BaseSet", dependencies = TRUE)'
 after_success:
     - Rscript -e 'covr::codecov()'
 
