+++ .github/workflows/R-CMD-check.yaml
@@ -13,14 +13,14 @@ jobs:
     steps:
       - uses: actions/checkout@v1
       - uses: r-lib/actions/setup-r@master
+      - name: Install system dependencies
+        if: runner.os == 'Linux'
+        run: sudo apt install libcurl4-openssl-dev
       - name: Install basic dependencies
         run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'BiocManager'))"
       - name: Install all dependencies
         run: Rscript -e "BiocManager::install('llrs/BaseSet', dependencies = TRUE)"
       - name: Check
         run: Rscript -e "rcmdcheck::rcmdcheck(args = '--no-manual', error_on = 'error')"
-      - name: Install system dependencies
-        if: runner.os == 'Linux'
-        run: sudo apt install libcurl4-openssl-dev
 
 
