+++ .github/workflows/R-CMD-check.yaml
@@ -13,13 +13,13 @@ jobs:
     steps:
       - uses: actions/checkout@v1
       - uses: r-lib/actions/setup-r@master
-      - name: Install system dependencies
+      - name: Install Linux dependencies
         if: runner.os == 'Linux'
         run: sudo apt install libcurl4-openssl-dev
       - name: Install basic dependencies
         run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'BiocManager'))"
       - name: Install all dependencies
-        run: Rscript -e "BiocManager::install('llrs/BaseSet', dependencies = TRUE)"
+        run: Rscript -e "BiocManager::install('llrs/BaseSet', dependencies = TRUE, build_vignettes = TRUE)"
       - name: Check
         run: Rscript -e "rcmdcheck::rcmdcheck(args = '--no-manual', error_on = 'error')"
 
