### 2020-02-14 7636a2f
Add link to Github Actions
Add also sessionInfo at the end of the builds
