### 2020-02-14 #359 @llrs
@annakrystalli I've been experimenting with [macOS builders](https://github.com/llrs/BaseSet/runs/445681741) and using `BiocManager::install('llrs/BaseSet', dependencies = TRUE, build_vignettes = TRUE)` it could install BaseSet and all its dependencies from Bioconductor without problems.
The test was on Catalina OS, CRAN checks the packages with El Captain, and you are using Mojave version. I'm not sure if I can find a CI with Mojave.
