+++ README.Rmd
@@ -33,7 +33,7 @@ It also allows to import from several formats used in the life science world. Li
 
 You can save information about the elements, sets and their relationship on the object itself. For instance origin of the set, categorical or numeric data associated with sets...
 
-Watch BaseSet working on the [examples](#Examples) below and in the vignettes. You can also find [related packages](#Related-packages) and the differences with BaseSet. If you have some questions or bugs [open an issue](https://github.com/llrs/BaseSet/issues) (remember of the [Code of Conduct](#Code-of-Conduct))
+Watch BaseSet working on the [examples](#Examples) below and in the vignettes. You can also find [related packages](#Related-packages) and the differences with BaseSet. If you have some questions or bugs [open an issue](https://github.com/llrs/BaseSet/issues) (remember the [Code of Conduct](#Code-of-Conduct))
 
 # Installation
 
@@ -177,7 +177,7 @@ There are several other packages related to sets, which partially overlap with B
  This package is focused on clustering of sets that are inside other sets and visualizations. However, BaseSet is focused on storing and manipulate sets including hierarchical sets.
  
  - [set6](https://cran.r-project.org/package=set6)
- This package implements different classes for different type of sets including fuzzy sets, conditional sets. However, it doesn't handle information associated to the elements or sets. 
+ This package implements different classes for different type of sets including fuzzy sets, conditional sets. However, it doesn't handle information associated to elements, sets or relationship. 
 
 # Code of Conduct {#CoC}
 
