### 2020-03-25 a01e3ec
Structural modifications to use Rcpp
