+++ R/BaseSet-package.R
@@ -1,0 +1,9 @@
+## usethis namespace: start
+#' @useDynLib BaseSet, .registration = TRUE
+#' @importFrom Rcpp sourceCpp
+## usethis namespace: end
+NULL
+
+#' @details
+#' It provides a class [`TidySet`] with methods to do operations with sets.
+"_PACKAGE"
