+++ R/basesets-package.R
@@ -1,3 +1,0 @@
-#' @details
-#' It provides a class [`TidySet`] with methods to do operations with sets.
-"_PACKAGE"
