### 2020-03-25 12e95ce
Start writing length functions in C++
