+++ src/RcppExports.cpp
@@ -1,0 +1,54 @@
+// Generated by using Rcpp::compileAttributes() -> do not edit by hand
+// Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
+
+#include <Rcpp.h>
+
+using namespace Rcpp;
+
+// prod2
+double prod2(NumericVector x);
+RcppExport SEXP _BaseSet_prod2(SEXP xSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericVector >::type x(xSEXP);
+    rcpp_result_gen = Rcpp::wrap(prod2(x));
+    return rcpp_result_gen;
+END_RCPP
+}
+// not_in
+NumericVector not_in(NumericVector p, NumericVector i);
+RcppExport SEXP _BaseSet_not_in(SEXP pSEXP, SEXP iSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericVector >::type p(pSEXP);
+    Rcpp::traits::input_parameter< NumericVector >::type i(iSEXP);
+    rcpp_result_gen = Rcpp::wrap(not_in(p, i));
+    return rcpp_result_gen;
+END_RCPP
+}
+// multiply_probabilities2
+NumericVector multiply_probabilities2(NumericVector p, NumericVector i);
+RcppExport SEXP _BaseSet_multiply_probabilities2(SEXP pSEXP, SEXP iSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericVector >::type p(pSEXP);
+    Rcpp::traits::input_parameter< NumericVector >::type i(iSEXP);
+    rcpp_result_gen = Rcpp::wrap(multiply_probabilities2(p, i));
+    return rcpp_result_gen;
+END_RCPP
+}
+
+static const R_CallMethodDef CallEntries[] = {
+    {"_BaseSet_prod2", (DL_FUNC) &_BaseSet_prod2, 1},
+    {"_BaseSet_not_in", (DL_FUNC) &_BaseSet_not_in, 2},
+    {"_BaseSet_multiply_probabilities2", (DL_FUNC) &_BaseSet_multiply_probabilities2, 2},
+    {NULL, NULL, 0}
+};
+
+RcppExport void R_init_BaseSet(DllInfo *dll) {
+    R_registerRoutines(dll, NULL, CallEntries, NULL, NULL);
+    R_useDynamicSymbols(dll, FALSE);
+}
