+++ R/RcppExports.R
@@ -1,0 +1,15 @@
+# Generated by using Rcpp::compileAttributes() -> do not edit by hand
+# Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
+
+prod2 <- function(x) {
+    .Call('_BaseSet_prod2', PACKAGE = 'BaseSet', x)
+}
+
+not_in <- function(p, i) {
+    .Call('_BaseSet_not_in', PACKAGE = 'BaseSet', p, i)
+}
+
+multiply_probabilities2 <- function(p, i) {
+    .Call('_BaseSet_multiply_probabilities2', PACKAGE = 'BaseSet', p, i)
+}
+
