+++ pkgdown.yml
@@ -1,8 +1,9 @@
 pandoc: '2.2'
-pkgdown: 1.4.1
+pkgdown: 1.5.1
 pkgdown_sha: ~
 articles:
   advanced: advanced.html
   basic: basic.html
   fuzzy: fuzzy.html
+last_built: 2020-04-09T17:10Z
 
