+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   advanced: advanced.html
   basic: basic.html
   fuzzy: fuzzy.html
-last_built: 2020-04-09T17:10Z
+last_built: 2020-04-10T17:09Z
 
