+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   advanced: advanced.html
   basic: basic.html
   fuzzy: fuzzy.html
-last_built: 2020-04-10T17:09Z
+last_built: 2020-04-11T17:10Z
 
