+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   advanced: advanced.html
   basic: basic.html
   fuzzy: fuzzy.html
-last_built: 2020-04-11T17:10Z
+last_built: 2020-04-12T17:12Z
 
