+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   advanced: advanced.html
   basic: basic.html
   fuzzy: fuzzy.html
-last_built: 2020-04-12T17:12Z
+last_built: 2020-04-13T17:12Z
 
