### 2020-06-03 #359 @llrs
@annakrystalli It's been a week since last message, and I haven't seen any plan announced here, on the website or on twitter. Is there a plan to start again looking for reviewers ?
PS: I am considering submitting directly to CRAN without waiting for a review from the volunteers of rOpenSci if it fits the [new scope](https://ropenscilabs.github.io/statistical-software-peer-review/scope.html#scope-categories).
