### 2020-06-03 #359 @llrs
Hi @annakrystalli Thanks for the quick reply, hope that everything is fine at your place.

I don't know how important is to wait for the second reviewer, maybe the first one can start without waiting for the second one? There are some authors of related packages that I would appreciate their feedback even if they have some conflict of interest against the package (I pasted their Github username): RaphaelS1,  kevinrue.
