### 2020-06-05 #359 @llrs
Many thanks in advance @arendsee and @j23414. As you can see on the isse you might find some difficulties installing the package. Let me know and I'll be happy to help with them.
