### 2020-06-06 #359 @arendsee
@llrs I installed the package from github and `devtools::check()` passed without unusual trouble (I'm on R4.0.0). So far so good!
