### 2020-06-07 #359 @j23414
@llrs Same here, I used `BiocManager::install`, and `library(BaseSet)` looks good on MacOS, R4.0.0. I'll clone the repo and do a `devtools::check`
