### 2020-06-17 #359 @arendsee
*Please check off boxes as applicable, and elaborate in comments below. Your
review is not limited to these topics, as described in the reviewer guide*

 None

- [x] As the reviewer I confirm that there are no [conflicts of
  interest](https://devguide.ropensci.org/policies.html#coi) for me to review
  this work (If you are unsure whether you are in conflict, please speak to
  your editor _before_ starting your review).

One of the vignettes should discuss the theory behind the fuzzy set operators
(see comments below).

##### For packages co-submitting to JOSS

- [x] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)

      and a reasonable range of inputs and conditions. All tests pass on the
      local machine.

Currently check fails on my system (R4.0.0 on Linux).

---

`BaseSet` is a well-designed and neatly written package that has clear value within the bioinformatics community and beyond. The functions are well documented and the test coverage is quite good. I think BaseSet should be admitted to rOpenSci once the issues I detail below are addressed. Most of them are minor. There are two major points:

 1. `check` currently does not pass on my system. I can build the vignettes and run the examples, so this is likely a minor dependency issue, or something, but it needs to be fixed.

 2. The theory behind the package (fuzzy set theory) needs much more explanation and justification. As it stands, I think most bioinformaticians and statisticians will misinterpret what the package is actually doing.

**Stylistic issues**

 * Indenting is inconsistent. Some functions indent 2 spaces and some 4.
 * remove empty lines from the ends of files (e.g., `add_relation.R`)
 * remove empty lines from the beginnings of files (e.g., `adjacency.R` and `head.R`)

**Line-by-line comments**

A quick scan of the code tells me it is generally clean and well organized. I read the vignettes and spot-checked a few files in more detail.  Below I've listed a few minor issues with file names and line numbers relative to commit 7d0f6da.

*spelling and grammar*
 * length.R:82 "... to simplify **tha** process for large objects .."
 * union.R:8; AllGenerics.R:303,325,348,368,388 **TidySets**", should be **TidySet**
 * basic.Rmd:56 "... for each element or set it can be **used added**:"
 * vignettes/basic.Rmd:220 "**to** where it is" should be "**of** where it is"
 * vignettes/fuzzy.Rmd:38 Missing period at end of sentence.
 * vignettes/fuzzy.Rmd:169/175 "type **do** we have" should be "type we have"

*miscellaneous*
 * length.R:199 remove `# browser()` line
 * subtract.R:15 remove `# browser()` line
 * GMT.R:4 Define the term GMT. I'm a bioinformatician and I am not familiar with the format, so I think adding a little detail would help. At minimum, you can expand the short description from "Import GMT files" to "Import GMT (Gene Matrix Transposed) files", this at least gives users something to google.
 * GMT.R:7 "Separation of the file" should be "GMT file field separator"
 * GMT.R:8,20  The documentation says "..." is passed to `strsplit`, but in the
   code "..." is passed to `readLines`
 * naming.R:57,74 `&` should be `&&`, since these is a scalar comparison

**(Optional) in the advanced vignette, replace humans with yeast**

The human genome, reactome, and GO dependencies of the advanced vignette take a very long time to download. Since one use case of a vignette is as a demo for a class or workshop, an hour long download is not ideal. If you want, you could replace the human data sets with smaller ones from yeast.

**Clarify annotations of "..."**

Below is a list of every "..." documentation string:

```
mutate.R:12:       #' @param ... The logical predicates in terms of the variables of the sets
power_set.R:8:     #' @param ... Other arguments passed down if possible
GMT.R:8:           #' @param ... Other arguments passed to strsplit
group.R:8:         #' @param ... A logical condition to subset some elements
complement.R:80:   #' @param ... Other arguments passed to either \code{\link{complement_set}} or
group_by.R:12:     #' @param ... The logical predicates in terms of the variables of the sets
list.R:17:         #' @param ... objects, possibly named (currently ignored).
select.R:14:       #' @param ... The name of the columns you want to keep, remove or rename.
droplevels.R:36:   #' @param ... Other arguments, currently ignored.
data_frame.R:22:   #' @param ... Other arguments currently ignored.
filter.R:13:       #' @param ... The logical predicates in terms of the variables of the sets
union.R:19:        #' @param ... Other arguments.
add.R:92:          #' @param ... Other arguments passed along.
cartesian.R:11:    #' @param ... Other arguments passed down if possible
AllGenerics.R:328: #' @param ... Other arguments.
AllGenerics.R:352: #' @param ... Other arguments.
AllGenerics.R:372: #' @param ... Other arguments.
AllGenerics.R:392: #' @param ... Other arguments.
AllGenerics.R:413: #' @param ... Other arguments.
size.R:5:          #' @param ... Other arguments to filter which size should be shown.
arrange.R:14:      #' @param ... Comma separated list of variables names or expressions
```

    Most of your documentation for "..." is fine. For example, "Comma separated list of variables names or expressions" tells me exactly what I need to know. Also, "Other arguments, currently ignored", is a perfectly acceptable.  However, "Other arguments", is not helpful and I would have to dig through the source code to figure out what parameters are supported (and I've had to do exactly this for a lot of packages).

**Document and test the handling of empty sets**

A `TidySet` is a set of elements

 * Empty sets:
   - `tidySet(list())`
   - `tidySet(list(a="A", b=character(0)))`

```
x <- list(a=c("A", "B"), b=character(0))
s <- tidySet(x)  # good - no errors are raised and the set looks right
elements(s)      # good
sets(s)          # debatable - the empty set disappeared
```

The "right" behaviour of `sets(s)` is debatable, but whatever you decide, the behavior of edge cases should be documented and covered in the test suite. I have no problem with the empty set disappearing, though it could cause problems with downstream code if the user assumes every input set will be listed in `sets(s)`.

In the `sets` documentation, I think the behavior would be clear if you just changed the Description to, "Given a TidySet, retrieve **all non-empty** sets or substitute them".

**Document and test the ordering of sets**

Should users assume sets are unordered? If so, you should explicitly say that sets are unordered and the order in which they are printed may change between implementations and should not be relied on in production code.

What happens when I add an element to a set? Is a new row added to the end of the table?

The `add_column` function seems a bit dangerous since it assumes an order. Just something to think about, you don't need to change anything.

**Fuzzy set math explanation**

Consider the following borrowed from the `fuzzy.Rmd` vignette:

```
> set.seed(4567)
> relations <- data.frame(sets = c(rep("A", 5), "B", "C"),
                          elements = c(letters[seq_len(6)], letters[6]),
                          fuzzy = runif(7))
> fuzzy_set <- tidySet(relations)
> fuzzy_set
  elements sets     fuzzy
1        a    A 0.2309186
2        b    A 0.7412554
3        c    A 0.1833834
4        d    A 0.4221682
5        e    A 0.5996399
6        f    B 0.2773313
7        f    C 0.7307312
> BaseSet::union(fuzzy_set, sets=c("C", "B"))
  elements sets     fuzzy
1        f  C∪B 0.7307312
```

When I first saw this result, I thought the math was simply wrong.  I would expect that the *probability* that `f` is in (C or B) would be greater than or equal to the probability that `f` is only in C. Why is it *exactly* equal?  Statistically, this would only be true if B were a subset of C.

After a bit of research, though, I found you are indeed using the standard operator definitions for fuzzy sets. If the fuzziness were the same as probability, then the union operator is a function that returns the lower bound of the probability of membership in the larger set. Do sets of genes expressed in different tissues follow the assumptions the union operator is making?

Fuzzy sets are based on [fuzzy logic](https://plato.stanford.edu/entries/logic-fuzzy).  In fuzzy logic, the 0 to 1 values are not probabilities, they are *truth values*. They are measures of the vagueness of a term not the than the chance of membership. You should describe the difference both in your README and the fuzzy vignette.

Most statisticians and bioinformaticians will not be familiar with fuzzy logic.  Also, most programs in the field deal with classification probabilities not fuzzy truth values. Although, it is possible that the values associated with tissue annotations are measures of the degree to which a gene is active in the specific tissue, in which case the fuzzy logic may apply.

So, you definitely need to describe the math and statistics behind the operators. Is there any prior research on the application of fuzzy sets in bioinformatics? If so, you might want to add some citations. You also need to justify the choice to use fuzzy set theory rather than more conventional statistics.
