### 2020-06-17 #359 @llrs
Many thanks for the fast, long and thoughtful review @arendsee! I will start fixing some of the issues you mentioned while I think about the theoretical question. Hope to reply them in a couple of days.
