### 2020-06-18 0406710
Remove double blank lines For #34
