### 2020-06-18 f911c6e
Remove double blank lines on tests for #34
