+++ tests/testthat/test-subtract.R
@@ -15,7 +15,6 @@ test_that("subtract works", {
   expect_equal(name_sets(b)[4], "(a∪b)∖c")
 })
 
-
 test_that("error", {
   TS <- tidySet(data.frame(elements = c("c", "c", "a", "b", "b"),
                 sets = c("A", "B", "B", "B", "C")))
