### 2020-06-18 54d6975
Remove empty blank line at the beginning of the document for #34
