### 2020-06-18 cfaa71b
Fix style, closes #34
