+++ R/group_by.R
@@ -14,12 +14,15 @@ dplyr::group_by
 #' @family methods
 #' @seealso dplyr \code{\link[dplyr]{group_by}} and \code{\link{activate}}
 #' @examples
-#' relations <- data.frame(sets = c(rep("a", 5), "b", rep("a2", 5), "b2"),
-#'                         elements = rep(letters[seq_len(6)], 2),
-#'                         fuzzy = runif(12))
+#' relations <- data.frame(
+#'     sets = c(rep("a", 5), "b", rep("a2", 5), "b2"),
+#'     elements = rep(letters[seq_len(6)], 2),
+#'     fuzzy = runif(12)
+#' )
 #' a <- tidySet(relations)
 #' elements(a) <- cbind(elements(a),
-#'                  type = c(rep("Gene", 4), rep("lncRNA", 2)))
+#'     type = c(rep("Gene", 4), rep("lncRNA", 2))
+#' )
 #' group_by(a, elements)
 #' @rdname group_by_
 #' @export
