+++ tests/testthat/test-toimplement.R
@@ -1,8 +1,10 @@
 context("test-head")
 
-relations <- data.frame(sets = c(rep("a", 5), "b"),
-                        elements = letters[seq_len(6)],
-                        fuzzy = runif(6))
+relations <- data.frame(
+    sets = c(rep("a", 5), "b"),
+    elements = letters[seq_len(6)],
+    fuzzy = runif(6)
+)
 a <- tidySet(relations = relations)
 
 test_that("head", {
