### 2020-06-18 e9bd93a
Fix miscellaneous code and documentation comments, closes #35
