### 2020-06-18 3930b5d
Check that a bug is no longer present, closes #43
