+++ tests/testthat/test-element_size.R
@@ -20,3 +20,16 @@ test_that("element_size works", {
     a1 <- add_elements(a, element = "g")
     expect_equal(element_size(a1, "g")$size, 0)
 })
+
+test_that("element_size and filter work well together", {
+    set.seed(4567)
+    relations <- data.frame(sets = c(rep("A", 5), "B", "C"),
+                            elements = c(letters[seq_len(6)], letters[6]),
+                            fuzzy = runif(7))
+    fuzzy_set <- tidySet(relations)
+    easy <- element_size(fuzzy_set, "f")
+    filtered <- fuzzy_set %>%
+        filter(sets %in% c("B", "C")) %>%
+        element_size(element = "f")
+    expect_equal(dim(easy), dim(filtered))
+})
