### 2020-06-18 e23bb1d
tidySet.list now accepts empty sets
Increased the checks on tidSet.data.frame
