### 2020-06-18 fb00b9f
tidySet.data.frame provides more informative errors
