### 2020-06-18 a3d7e47
Change a test to allow keeping elements but with no sets and thus relations
