### 2020-06-18 9b148ed
Check that list with no elements are correctly considered as sets.
