### 2020-06-18 1e68c66
Slight improvement of the documentation
And update of the newest roxygen2
