+++ man/name_sets-set.Rd
@@ -18,9 +18,11 @@ A \code{TidySet} object.
 Rename sets.
 }
 \examples{
-relations <- data.frame(sets = c(rep("a", 5), "b"),
-                        elements = letters[seq_len(6)],
-                        fuzzy = runif(6))
+relations <- data.frame(
+    sets = c(rep("a", 5), "b"),
+    elements = letters[seq_len(6)],
+    fuzzy = runif(6)
+)
 a <- tidySet(relations)
 a
 name_sets(a) <- LETTERS[1:2]
