### 2020-06-18 a3be21c
Check all the fulty tests
