### 2020-06-18 f995cd1
Fix warning on check and improve documentation
