+++ README.Rmd
@@ -37,7 +37,7 @@ Watch BaseSet working on the [examples](#Examples) below and in the vignettes. Y
 
 # Installation
 
-The package depends on some packages from Bioconductor. In order to install some of its dependencies you'll need first to install BiocManager:
+The package depends on some packages from Bioconductor. In order to install some of its dependencies you'll need first to install `{BiocManager}`:
 
 ```{r dep, eval = FALSE}
 if (!require("BiocManager")) {
@@ -167,16 +167,16 @@ There are several other packages related to sets, which partially overlap with B
  - [sets]( https://CRAN.R-project.org/package=sets)  
  Implements a more generalized approach, that can store functions or lists as an element of a set (while BaseSet only allows to store a character or factor), but it is harder to operate in a tidy/long way. Also the operations of intersection and union need to happen between two different objects, while a single TidySet object (the class implemented in BaseSet) can store one or thousands of sets.
 
- - [GSEABase](https://bioconductor.org/packages/GSEABase)  
+ - [`{GSEABase}`](https://bioconductor.org/packages/GSEABase)  
  Implements a class to store sets and related information, but it doesn't allow to store fuzzy sets and it is also quite slow as it creates several classes for annotating each set. 
   
- - [BiocSets](https://bioconductor.org/packages/BiocSets)  
+ - [`{BiocSets}`](https://bioconductor.org/packages/BiocSets)  
  Implements a tidy class for sets but does not handle fuzzy sets. It also has less functionality to operate with sets, like power sets and cartesian product. BiocSets was influenced by the development of this package. 
 
- - [hierarchicalSets](https://CRAN.R-project.org/package=hierarchicalSets)  
+ - [`{hierarchicalSets}`](https://CRAN.R-project.org/package=hierarchicalSets)  
  This package is focused on clustering of sets that are inside other sets and visualizations. However, BaseSet is focused on storing and manipulate sets including hierarchical sets.
  
- - [set6](https://cran.r-project.org/package=set6)
+ - [`{set6}`](https://cran.r-project.org/package=set6)
  This package implements different classes for different type of sets including fuzzy sets, conditional sets. However, it doesn't handle information associated to elements, sets or relationship. 
 
 # Code of Conduct {#CoC}
