### 2020-06-18 d1ecdaf
Fix the documentation for the new argument
