### 2020-06-19 3096658
Review all the ... arguments
Also corrected some errors and fixed an error about the order of printing and of the TidySet
