+++ tests/testthat/test-arrange.R
@@ -21,3 +21,16 @@ test_that("arrange works", {
     b <- arrange(a, desc(sets))
     expect_equal(as.character(relations(b)[1, "sets"]), "b2")
 })
+
+test_that("check arrange by fuzzy values", {
+    set.seed(4987951)
+    r <- data.frame(
+        sets = c(rep("A", 5), "B", rep("A2", 5), "B2"),
+        elements = rep(letters[seq_len(6)], 2),
+        fuzzy = runif(12),
+        type = c(rep("Gene", 2), rep("Protein", 2), rep("lncRNA", 2))
+    )
+    TS <- tidySet(r)
+    s <- arrange(TS, fuzzy)
+    expect_equal(order(as.data.frame(s)$fuzzy), seq_len(12))
+})
