### 2020-06-20 ea0c3b6
Allow to pull to use unquoted symbols
Increase tests of pull
