### 2020-06-22 35ed8f7
Explicitly mention the empty sets
