### 2020-06-22 #359 @llrs
I updated the package following your comments @arendsee. See below for some questions and comments that I still need to address.

**Review Comments**

- [x] `check` fixed. Now with the latest version it should run on `dplyr` > 1.0.0

- [ ] I'm not sure this package is the right place to justify the use of fuzzy set theory on one area (bioinformatics). I agree that if I only show the capabilities of the package it won't be compelling enough, that's one of the reason why I created the long advanced vignette. I will explain better my reasoning of the fuzzy logic implemented on the package in a new section of the README on the fuzzy vignette to explain what is the fuzzy set logic, how it is implemented and differences with the typical sets.

**Stylistic issues**

 - [x] Fixed using `styler`

**Line-by-line comments**

 - [x] Spelling fixed, grammar not sure (I'm not a native English speaker)
 - [x] Miscellaneous fixed

**(Optional) in the advanced vignette, replace humans with yeast**

- [ ] I don't expect all users to install the packages required for the vignette, on this regard I followed Bioconductor's guidelines to provide a representative vignette of what can be done with the package, not only how it works. I think that most people work with human data so I wanted to provide compelling examples that at the same time they could be familiar and have already the packages installed. However if you point me to other yeast data set I will check it.

**Clarify annotations of "..."**

- [x] Corrected, and agree on the comment, they need to be clear enough for the user. I removed some `...` from methods, others I got them to use like in `union`. In any case the documentation was updated to explain and in some cases the examples were extended to show what they do.

**Document and test the handling of empty sets**

- [x] That was a corner case I missed I improved the creation of sets sets from list with empty characters. Empty sets are allowed but not shown when printing (documentation was updated to explain this). However, I disabled the creation of an empty TidySet from a list (no elements and no sets). Not sure if it is worth to support the creation of an empty object to fill it later. I think there are enough methods to create a TidySet when needed. Added three more tests for these cases.

**Document and test the ordering of sets**

 - [x] Sets are considered unordered unless you arrange them on a specific order. The order of printed sets may change if there are modification on the relations table. (There was an inconsistency between the table shown and the order it was stored, this was now fixed)

When a new element is added to a set via `add_relation`, first it is added to the end of the elements table, then it is added on the relations table at the end too.

The `add_column` adds columns to the existing table in the given order. It can be dangerous if the expected order is not meet, but I don't think it will be used much, as there are better alternatives to add a column and set some values on just some rows such as using `mutate(TidySet, column = case_when())`.

I do not know enough of ordered sets to know if I should provide specific support to them, or how are they usually used.

**Fuzzy set math explanation**

In this case both sets B and C are 1 element size, both of them have the `f` element. So taking the higher bound of the common values makes the probability of `f` equal to C. Note that the union does not calculate the probability that `f` is on C or B, but which is the probability of the common elements of C and B. To know the probability of `f` to be in C or B we need to use:

```r
element_size(fuzzy_set, "f")
#   elements size probability
# 6        f    0   0.1945921
# 7        f    1   0.6027532
# 8        f    2   0.2026547
```

Here we can see that there is a 0.6 probability that f is in either of them. (If `f` would be in other sets we need to first filter to only those: `fuzzy_set %>% filter(sets %in% c("B", "C")) %>% element_size(element = "f")`, in this case the result is the same.)

The  standard operator definitions for fuzzy are defined in the `union` argument fun, which is set to "max". However, then the definition of the common operation is not the same as it is implemented as you define it as "lower bound of the probability of membership in the larger set". In this case where both of them have the same number of elements involved, which should be picked? This could be a corner case, but as you explained they must be well covered and documented. In addition to that how is the union the "lower bound of the probability", wouldn't it be the higher bound of the probability?

Following your definition what would be the expected result of the union of B and C on the following example? With the previous seed, on this example B has more elements than C but has lower bound of *truth value*(?):

```r
relations <- data.frame(sets = c(rep("A", 5), "B", "B", "C"),
                          elements = c(letters[seq_len(7)], letters[6]),
                          fuzzy = runif(8))
fuzzy_set <- tidySet(relations)
fuzzy_set
```

I'm not sure I understand the assumptions of the union operator. But the purpose of this package is to test if fuzzy methods are helpful on sets and pathways tests. I do not know before hand if the fuzzy logic can explain better any system than other methodologies.

The theoretical background of fuzzy logic is broad and beyond my current comprehension. However, on the linked page there is an example (section 9) on how such a systems arise that is pertinent:

> Voting semantics is based on the idea that different agents (voters) may coherently judge the same proposition differently. The proportion of agents that accept a proposition φ as true may be seen as a truth value. Without further restrictions this does not lead to a truth functional semantics, but rather to an assignment of probabilities to propositions. But if one assigns a fixed level of skepticism to each agent and imposes some natural conditions that keep the judgments on logically complex statements consistent with those levels, then one can recover min, max, and 1−x as truth functions

On the case of pathways and genes there are several voters (scientists or algorithms) that provide their truth value of the association of a gene in a pathway. This is usually expressed as probabilities "I'm 80% sure that X gene is on Y pathway" or with evidence codes in the case of ontologies like gene ontology (i.e. [IEA](http://geneontology.org/docs/guide-go-evidence-codes/). This is why I don't make a distinction between vagueness and the probability of an association.

I understand the lack of exposure of this field on the bioinformatics, statistics and experimental fields with this logic. This lack of understanding can slow the adoption of the package but at the same time is one of the strengths as it provides with a new method not used much previously on the field. There are other experimental methods to measure the  degree to which a gene is active, mainly sequencing, however then the problem is distinguish between tissues and between cell types specific expression. I build a webpage where I collected an example of the application of BaseSet on this kind of research: https://llrs.github.io/BaseSet_scRNAseq/AUCell.html

- [ ] There are a couple of articles where fuzzy set methods are used, specially on the area of ontology assignment and comparison. I will provide them as examples of previous usage of these methods. However I haven't cited them because I find them hard to follow and I'm not sure I understand well how they apply the fuzzy logic on their respective topics.
