### 2020-06-22 #359 @llrs
Hi @annakrystalli, the changes are already pushed to master. I could create a new branch and move all the commits there but I hope it doesn't causes problems.

If you @j23414 are using `pkgreviewer` I think you'll be able to review the package at the same commit that @arendsee reviewed:

```r
pkgreviewr::pkgreview_create(pkg_repo = "llrs/BaseSet@7d0f6da", ...)
```

Otherwise let me know and I'll fix this on my end.
