### 2020-06-22 #359 @j23414
Hi @llrs, I'll continue my review with your changed version. I wrote a script on my end when I saw you had new commits (automatically pulls new version and run any Rmd files) but thanks for the `pkgreview_create` command.

I'm aiming to post by Friday.
