### 2020-06-22 #359 @arendsee
Thanks for the response @llrs . I'm happy with all the changes. As for fuzzy theory, I just want to be sure that the vignettes document what the operators are doing, especially where the fuzzy operators deviate from the more familiar probabilistic operators. Specifically, in probability theory, P(A or B) = P(A) + P(B) - P(A and B). In contrast, the fuzzy operator F(A or B) = max(F(A), F(B)). Let's wait for @j23414 to finish her review then we can talk about it.
