### 2020-06-22 #359 @llrs
Ok, let's wait. I'm not sure I understand your example, what are A and B on your example? Elements or sets?
