### 2020-06-22 #359 @arendsee
P(A or B) is the probability that A or B is true. In this context, it is the probability that an element is in set A or set B.
