### 2020-06-25 #359 @j23414
None

  <details>	<summary>click to expand</summary>

  > The goal of BaseSet is to facilitate working with sets in an
  efficient way. The package implements methods to work on sets, doing
  > intersection, union, complementary, power sets, cartesian product and
  > other set operations in a tidy way. Both for classical and fuzzy
  > sets.

  <details> <summary> </summary>

  > The package depends on some packages from Bioconductor. In order   to install some of its dependencies you’ll need first to install BiocManager:

  ```
  if (!require("BiocManager")) {
    install.packages("BiocManager")
  }
  ```

  > You can install the latest version of BaseSet from Github with:

  ```
  BiocManager::install("llrs/BaseSet",
                       dependencies = TRUE, build_vignettes = TRUE, force = TRUE)
  ```

    <details> <summary>Minor edit</summary>

    Rmd files in `vignettes` folder were built without any R runtime errors. Since I've had issues getting `devtools::vignette` to run for other R packages in the past, I am ignoring the `no vignettes found` message.

    ```
    devtools::vignette(package = "BaseSet")
    #> no vignettes found

    # generate the HTML file for each Rmd in vignettes folder
    rmarkdown::render("vignettes/basic.Rmd", "html_document")
    rmarkdown::render("vignettes/advanced.Rmd", "html_document")
    rmarkdown::render("vignettes/fuzzy.Rmd", "html_document")

    # opened HTML files and didn't see any R runtime errors
    ```

    **Minor edit:** In `vignette/fuzzy.Rmd` there's a line 'This vignettes supposes that you already read the “About Set” vignette' ... I think you mean "About BaseSet" generated from `basic.Rmd`.

    <details>	<summary> </summary>

    There are 187 exported functions, so I only skimmed the `R/*.R` files and checked for formatting and documentation. Nothing stood out as drastically wrong.

    ```
    exports <- ls("package:BaseSet")
    length(exports)
    #> 187
    ```

    Help worked fine in RStudio

    ```
    help(package = "BaseSet")
    # clicked through several pages
    ```

    <details><summary>Most examples ran, two examples threw errors</summary>

    ```
    # Clicked through some of the help pages in RStudio
    help(package = "BaseSet")
    ```

    Ended up pulling all the examples using bash, there's probably a simpler way to do this

    *bash terminal*

    ```
    # Examples in README.md ran fine.

    # Pull all @examples from BaseSet/R/*.R files
    cat BaseSet/R/*.R |\
    awk "/example/{flag=1}/^[^#]/{flag=0}flag" |\
    sed "s/^#' //g" |\
    sed "s/@examples/############# examples/g" |\
    sed "s/^@/# /g" > BaseSet_Examples.Rmd

    # Add a standard Rmd header, run all examples
    open -a Rstudio BaseSet_Examples.Rmd
    ```

    **Two examples threw errors:**

    ```
    # File: set.R; setMethod("sets<-"
    sets(TS) <- data.frame(sets = c("B", "A"))
    #> Error in validObject(object) :
    #> invalid class “TidySet” object: All sets must be described on their table.

    # File: set.R; Function: replace_sets()
    TS2 <- replace_sets(TS, data.frame(sets = c("A", "B", "C")))
    #> Error in validObject(object) :
    #> invalid class “TidySet” object: All sets must be described on their table.
    ```

    <details><summary> </summary>

    In the `CODE_OF_CONDUCT.md` file

  <details> <summary> </summary>

  Installed fine on MacOS 10.13.6 (R v4.0.0) after updating BiocManager and a few R packages.

  ```
  # (1) README install
  if (!require("BiocManager")) {
    install.packages("BiocManager")
  }
  BiocManager::install("llrs/BaseSet",
                   dependencies = TRUE,
                   build_vignettes = TRUE,
                   force = TRUE)

  # (2) Installation methods via pkgReviewr pipeline
  devtools::install("~/Desktop/BaseSet", dependencies = T, build_vignettes = T)
  devtools::install_github("llrs/BaseSet", dependencies = T, build_vignettes = T)
  #>  Skipping 6 packages not available: BiocStyle, GO.db, GSEABase, org.Hs.eg.db, reactome.db, Biobase
  ```

  <details><summary>`set_size` was having weird behavior</summary>

  Created a toy example of E. coli.

  * elements = genes
  * sets = pathways

  ```
  ecoli_genes <- paste("gene", c(1:4401), sep = ":") # 4401 E. coli genes
  ecoli_sets <-
  list(GL = ecoli_genes[1:4000],            # Glycolysis
       CF = ecoli_genes[3000:4401],         # Carbon fixation
       AF = ecoli_genes[c(1:2000) * 2]) %>% # Ability to fly
  tidySet(.)
  head(ecoli_sets, n = 3)
  #>   elements sets fuzzy
  #> 1   gene:1   GL     1
  #> 2   gene:2   GL     1
  #> 3   gene:3   GL     1
  ```

  Check union/intersect/negation functions - worked as expected

  ```
  ecoli_sets %>%
  union(., set = c("GL", "CF")) %>%
  #  intersection(., set = c("GL", "CL", "AF")) %>%    # could be called intersect
  #  subtract(., set_in = "GL", not_in = "AF") %>%
  head(., n = 3)
  #>   elements  sets fuzzy
  #> 1   gene:1 GL∪CF     1
  #> 2   gene:2 GL∪CF     1
  #> 3   gene:3 GL∪CF     1
  ```

  (Optional change) Could change parameter `set = c("GL", "CL"` to the alternative `sets = ...` and match the tidySet header `elements`, `sets`,  `fuzzy`.

  ```
  set_size(ecoli_sets)
  #>   sets size probability
  #> 1   GL 4000           1
  #> 2   CF 1402           1
  #> 3   AF 2000           1
  ```

  **Fuzzy sets - some weird behavior**

  Noticed some weird behavior in `set_size` function.

  ```
  # For the fuzzy test, use a smaller dataset
  (ecoli_sets <-
  list(GL = ecoli_genes[1:2],
       CF = ecoli_genes[2:4]) %>%
  tidySet(.))
  #>   elements sets fuzzy
  #> 1   gene:1   GL     1
  #> 2   gene:2   GL     1
  #> 3   gene:2   CF     1
  #> 4   gene:3   CF     1
  #> 5   gene:4   CF     1

  set_size(ecoli_sets)  # Not fuzzy size
  #>   sets size probability
  #> 1   GL    2           1
  #> 2   CF    3           1

  (ecoli_sets %>%
    mutate(
      fuzzy = case_when(sets == "GL" ~ 0.2,    # Add fuzzy values
                        sets == "CF" ~ 0.8)
    ))
  #>   elements sets fuzzy
  #> 1   gene:1   GL   0.2
  #> 2   gene:2   GL   0.2
  #> 3   gene:2   CF   0.8
  #> 4   gene:3   CF   0.8
  #> 5   gene:4   CF   0.8

  set_size(ecoli_sets)  # Fuzzy size
  #>   sets size probability
  #> 1   GL    2           1
  #> 2   CF    3           1
  ```

  Wait ... the fuzzy size shouldn't match the not-fuzzy size ... I'll create the same fuzzy set following the instructions in the README.md and check its size.

  (ecoli_sets <- data.frame(sets = c(rep("GL", 2), rep("CL", 3)),
                        elements = c(ecoli_genes[1:2], ecoli_genes[2:4]),
                        fuzzy = c(rep(0.2, 2), rep(0.8, 3))) %>%
  tidySet(.))
  #>   elements sets fuzzy
  #> 1   gene:1   GL   0.2     # <=
  #> 2   gene:2   GL   0.2     # <=
  #> 3   gene:2   CL   0.8
  #> 4   gene:3   CL   0.8
  #> 5   gene:4   CL   0.8

  set_size(ecoli_sets)   # Fuzzy size, again... this is what I expected
  #>   sets size probability
  #> 1   CL    0       0.008
  #> 2   CL    1       0.096
  #> 3   CL    2       0.384
  #> 4   CL    3       0.512
  #> 5   GL    0       0.640
  #> 6   GL    1       0.320
  #> 7   GL    2       0.040   # Note: 0.040 = 0.2 * 0.2 from above
  ```

  Any explanation appreciated.

  Looking at the table, the fuzzy values are being treated as probabilities. Last line of the table shows that GL (glycolysis) has 2 genes with a probability of 0.04 which is equal to 0.2 x 0.2. So the biological interpretation is that "gene:1" is classified as “Glycolysis” 0.2 (20%) of the time? And ditto for "gene:2"? Ergo the most likely outcome is that “Glycolysis” has 0 genes with a probability of 0.64 and “Carbon fixation” has 3 genes with probability of 0.512. Hmm, I'll think on the interpretation of that probability column, but seems to matches the example in the Fuzzy vignette. Searched for papers on fuzzy-set size/cardinality that used probabilities, but all searches redirected me to a sum of fuzzy values (0.2 + 0.2). Adding a description on how the probability column is calculated would be sufficient documentation to avoid confusion.

  <details><summary></summary>

  `Advanced.Rmd` used BaseSet on gene ontology and pathway data and finished in reasonable time.

   	<details> <summary> </summary>

	Ran the following automated tests. Only included any failed, warning, skipped and relevant lines of the output. Most checks passed.

    ```
    pkg_dir="~/Desktop/BaseSet"
    devtools::check(pkg_dir)
    #> 0 errors | 0 warnings | 0 notes
    devtools::test(pkg_dir)
    #> Testing BaseSet
    #> ✓ |  OK F W S | Context
    #> x |   0 1 1   | test-print
    #> test-print.R:9: warning: print works
    #> cannot open compressed file '~/Desktop/BaseSet-review/.Rproj.user/shared/notebooks/543CF448-index/1/3714FEAA6C457ACE/c0tuarduqr19d_t/_rs_rdf_3674270fe467.rdf', probable reason 'No such file or directory'
    #> test-print.R:9: error: print works
    #> cannot open the connection
    #> Backtrace:
    #> 1. testthat::expect_output(show(a), "elements\\s+sets\\s+fuzzy") tests/testthat/test-print.R:9:4
    #> 11. BaseSet::show(a)
    #> 13. print.data.frame(o)
    #> 14. overridePrint(o$x, o$options, o$className, o$nRow, o$nCol)
    #> 15. base::save(x, options, file = output)
    #> 16. base::gzfile(file, "wb")
    #> ✓ |  12     1 | test-tidy [3.1 s]
    #> test-tidy.R:41: skip: tidy works in GeneColorSet
    #> Reason: experimentData is not imported
    #> ====== Results
    #> OK:       381
    #> Failed:   1
    #> Warnings: 1
    #> Skipped:  1
    ```

    The `test-print failed` seems to be file permission problems, so I'm assuming it's a problem specific to my laptop. So I'm going to ignore it.

    Not sure why we're skipping the `test-tidy` or why we're not importing `experimentData`. Including output here in case it matters to you.

    ```
    goodpractice::gp(pkg_dir)
    #> GP BaseSet
    #> It is good practice to
    #> ✖ write unit tests for all functions, and all package code in
    #> general. 88% of code lines are covered by test cases.
    #> R/add_relation.R:22:NA
    #> R/add.R:55:NA
    #> R/add.R:56:NA
    #> R/add.R:57:NA
    #> R/add.R:60:NA
    #> ... and 143 more lines
    #>
    #>  ✖ avoid long code lines, it is bad for readability. Also, many
    #>  people prefer editor windows that are about 80 characters wide. Try
    #>  make your lines shorter than 80 characters
    #>
    #>  R/AllClasses.R:7:1
    #>  R/AllGenerics.R:90:1
    #>  R/AllGenerics.R:126:1
    #>  R/AllGenerics.R:143:1
    #>  R/GMT.R:15:1
    #>  ... and 11 more lines
    ```

    It's good to have high test coverage, but as long as you test major functions I guess this is fine.

    The "longer than 80 character width" won't break your code, so this passes for me. I agree that "80 char widths" is good practice in any programming language.

    ```
    devtools::spell_check(pkg_dir)   # passed with reasonable spelling
    ```

- [x] **Packaging guidelines**: The package conforms to the [rOpenSci packaging guidelines](https://devguide.ropensci.org/building.html)

  <details><summary></summary>

  Did a quick code-check

  **bash console**

  ```
  grep "print(" BaseSet/R/*.R     # Passed
  grep "cat(" BaseSet/R/*.R       # Passed
  grep "->" BaseSet/R/*.R         # Passed

  ```

---

BaseSet mostly passes my automated checks. Just some minor edits in the vignettes and the examples. (Expand sections above.) I was getting some strange results with the `set_size` method (see the Functionality section).

Creating sets that are [Tidyverse](https://tidyverse.tidyverse.org/articles/paper.html)-friendly makes sense. The Classical-set operators seem clear enough and matches what I've seen in the past. The advanced vignette is mainly focused on classical-sets.

**Fuzzy-set operators**

For fuzzy-sets, alternative definitions for union/intersect exist. The probability function mentioned by @arendsee is also mentioned in Fuzzy Set Theory (Smithson, 2006) ([link to chapter, see page 10, below Table 2.3](https://us.sagepub.com/sites/default/files/upm-assets/11134_book_item_11134.pdf)) as a product operator, which is used extensively in probability theory (statistics). Documenting the fuzzy-set operators as min/max/etc with a link to a paper (Zadeh?) should be sufficient to avoid confusion.

The [sets](https://cran.r-project.org/web/packages/sets/sets.pdf) package (see `?sets::fuzzy_logic` help file) lets the user choose `Zadeh` (min/max) or `product` forms of fuzzy logic.

Just a side-note (not a change request): Instead of "probability", fuzzy-set papers seem to use the term  [membership](https://en.wikipedia.org/wiki/Membership_function_(mathematics)) or [membership value](http://www.dma.fi.upm.es/recursos/aplicaciones/logica_borrosa/web/fuzzy_inferencia/funpert_en.htm#:~:text=Definition%3A%20a%20membership%20function%20for,to%20the%20fuzzy%20set%20A.). The Fuzzy vignette example was helpful, classifying cells to different cell types along with a fuzzy score, which sounds like a "confidence score" or "[confidence level](https://en.wikipedia.org/wiki/Confidence_interval)".

**Other vignette comments**

The `vignette/advanced.Rmd` mainly focused on classical sets (not fuzzy), where elements=genes and sets=GO terms or pathways. While I appreciated seeing a larger dataset, I could have used more discussion on why the BaseSet-generated results were important. In figure one, gene ontology is split into BP, CC, and MF. Spelling out their names as "biological pathway", "cellular component", and "molecular function" helps readability. A key takeway from that first figure could be that BP "biological pathway" terms tend to be assigned using IBA evidence which stands for "inferred from biological aspect of ancestor". After going through `basic.Rmd` and `fuzzy.Rmd`, I was expecting to see different evidence codes assigned different fuzzy values (measure of confidence), and the maximal value used to assign a gene to a particular pathway. You don't need to add the fuzzy values (just my response to the reading). Counting up number of genes per pathway or per gene ontology term was fine. Basic and Fuzzy vignettes were fine.

**My R Session Info**

<details><summary>sessionInfo()</summary>

```
#> R version 4.0.0 (2020-04-24)
#> Platform: x86_64-apple-darwin17.0 (64-bit)
#> Running under: macOS High Sierra 10.13.6
#>
#> Matrix products: default
#> BLAS:   /Library/Frameworks/R.framework/Versions/4.0/Resources/lib/libRblas.dylib
#> LAPACK: /Library/Frameworks/R.framework/Versions/4.0/Resources/lib/libRlapack.dylib
#>
#> locale:
#> [1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8
#>
#> attached base packages:
#> [1] parallel  stats4    stats     graphics  grDevices utils     datasets
#> [8] methods   base
#>
#> other attached packages:
#>  [1] GO.db_3.11.4         org.Hs.eg.db_3.11.4  AnnotationDbi_1.50.0
#>  [4] IRanges_2.22.2       S4Vectors_0.26.1     Biobase_2.48.0
#>  [7] BiocGenerics_0.34.0  BaseSet_0.0.11       magrittr_1.5
#> [10] forcats_0.5.0        stringr_1.4.0        dplyr_1.0.0
#> [13] purrr_0.3.4          readr_1.3.1          tidyr_1.1.0
#> [16] tibble_3.0.1         ggplot2_3.3.2        tidyverse_1.3.0
#>
#> loaded via a namespace (and not attached):
#>  [1] Rcpp_1.0.4.6     lubridate_1.7.9  lattice_0.20-41  assertthat_0.2.1
#>  [5] digest_0.6.25    R6_2.4.1         cellranger_1.1.0 backports_1.1.8
#>  [9] reprex_0.3.0     RSQLite_2.2.0    evaluate_0.14    httr_1.4.1
#> [13] pillar_1.4.4     rlang_0.4.6      readxl_1.3.1     rstudioapi_0.11
#> [17] blob_1.2.1       rmarkdown_2.3    bit_1.1-15.2     munsell_0.5.0
#> [21] broom_0.5.6      compiler_4.0.0   modelr_0.1.8     xfun_0.15
#> [25] pkgconfig_2.0.3  htmltools_0.5.0  tidyselect_1.1.0 fansi_0.4.1
#> [29] crayon_1.3.4     dbplyr_1.4.4     withr_2.2.0      grid_4.0.0
#> [33] nlme_3.1-148     jsonlite_1.6.1   gtable_0.3.0     lifecycle_0.2.0
#> [37] DBI_1.1.0        scales_1.1.1     cli_2.0.2        stringi_1.4.6
#> [41] fs_1.4.1         xml2_1.3.2       ellipsis_0.3.1   generics_0.0.2
#> [45] vctrs_0.3.1      tools_4.0.0      bit64_0.9-7      glue_1.4.1
#> [49] hms_0.5.3        yaml_2.2.1       colorspace_1.4-1 rvest_0.3.5
#> [53] memoise_1.1.0    knitr_1.29       haven_2.3.1
```
