### 2020-06-26 #359 @llrs
Thanks @j23414 for your detailed review and apologies for modifying the package while you was trying to review it. Some of the modifications I did seem the cause of some of the errors you saw. I'll think more carefully about them and fix them.
Many thanks for pointing to unclear explanations and your comments on the fuzzy-set operators and the vignettes.
