### 2020-06-26 #359 @arendsee
The following statement in the README is not right: "On fuzzy sets, elements have a probability to belong to a set".

In a probabilistic framework, every element is in a set or not in it. We are uncertain which sets an element is in, but the reality is still binary. In a fuzzy framework, the reality is vague. For example, say we are interested in which genes are highly expressed in a tissue. We could transform RNA-seq data for expression levels to a 0-1 scale, where 1 indicates the gene is very highly expressed, 0.6 is fairly highly expressed, 0.3 is weakly expressed, etc. The reality here is not binary.

If you use fuzzy methods on probabilistic data, as you do in the README and the fuzzy vignette, the results will be wrong except in very special cases. The probability that a particular element is in A or B is `P(A) + P(B) - P(A and B)`. This is a law of probability. Application of the fuzzy operator to probabilistic terms would imply that `P(A or B) = max(P(A), P(B))`, which is correct only when `P(B) = P(A and B)` or `P(A) = P(A and B)`. In other words, it is true only when one set contains the other set.

As @j23414 mentioned, the probability function (assuming independence) can be defined to work over fuzzy values (Smithson, 2006). Take the statement, "I am tall OR strong". Now say my tall stat is 0.3 and my strong stat is 0.6. The truth value of this statement is 0.6, using the max operator. This intuitively seems reasonable. Similarly, it makes intuitive sense that the statement, "I am tall AND strong" has a truth value of 0.3. We can do similar calculations using the probability function where, assuming independence, we get `0.6 + 0.3 - 0.3 * 0.6 = 0.72`. This is a different truth value because we've redefined the fuzzy space by using a different operator. This different definition of fuzziness has some nice properties, so we may choose to use it for certain applications.

However, we don't get to define our own laws of probability. If we are given probabilistic data, then we have to use probabilistic operators. We can't freely change the data from membership probabilities to fuzzy truth values; these mean different things.

These distinctions need to be VERY clear in the README and vignette. If they aren't clear, a careless user could easily assume the operators are probabilistic and get completely wrong results.
