### 2020-06-29 #359 @llrs
@arendsee Perhaps that sentence on the README is not suited for some logics or types of fuzzy sets. I am struggling to introduce the topic for people not familiar with fuzzy sets and at the same time support and explain what the package can do to more expert users. I'll carefully review the wording around the subject.

I agree that it is different to work with membership functions (Gene A 0.3 RNA levels means low, 0.6 means highly expressed) than with probabilities (Gene A is on 60% of cases highly expressed, whatever its value is). However, assigning a set from a measure is, nowadays, out of the scope of the package, this is either provided by the source of data or something the user should do.

About probabilities of A or B, it is still unclear to me over what do you calculate that probability or when would you like to do this. When I'll post a longer reply I'll provide some examples and questions about this, as I tend to understand something better with a practical question or exercise. On your example `tall OR strong`, you don't use your previous definition that "the probability that an element is in set A or set B". What other elements are present on the height and strength set (and what are their probabilities)? And why do you use height and strength stat as probability or truth value? They are different things, that's why membership functions are used, or are they already the fuzzy values?

While users should not use different logics on the same data the package aims to allow them to use whatever logic they want according to the knowledge they have about the data. But I can't guide them or provide feedback.

The fuzzy set theory extends the probability framework as explained on [a source](http://www.cs.cmu.edu/Groups/AI/html/faqs/ai/fuzzy/part1/faq-doc-10.html) @j23414 provided:

>... fuzzy researchers have gone to great pains to distance
themselves from probability. But in so doing, many of them have lost track
of another point, which is that the converse DOES hold: all probability
distributions are fuzzy sets! As fuzzy sets and logic generalize Boolean
sets and logic, they also generalize probability.

----

Hope to correct the points raised by @j23414 and post about the set theory and fuzzy-set operators that both reviewers had questions by the end of the week.
