### 2020-06-29 #359 @j23414
Hmm, that quoted source is new to me (Bezdek, 1993)... I provided links to the set theory chapter ([Smithson 2006](https://us.sagepub.com/sites/default/files/upm-assets/11134_book_item_11134.pdf)), membership value definition, and confidence value definition.

No matter : ) thanks for sharing the source! There definitely seem to be some murkiness to when and where is appropriate to use Fuzzy Sets.

For @arendsee's `tall OR strong` example, he seems to be illustrating the different calculation and final value in:

Case 1) if `tall` (0.3) and `strong` (0.6) are treated as fuzzy truth values:
then `tall OR strong` = max (tall, strong) = max(0.3, 0.6) = 0.6

Case 2) if `tall` (0.3) and `strong` (0.6) are treated as probabilities (and assume independence): **
then `tall OR strong` = P(tall) + P(strong) - P(tall and strong) = 0.3 + 0.6 - (0.3 x 0.6) = 0.72

**You can also think of this as a `tall-ness gene` where you are "more likely to be tall, but not guaranteed to be tall"

The union of probabilities (Case 2) is usually explained and illustrated as a venn diagram ([link with explanation here](https://courses.lumenlearning.com/boundless-statistics/chapter/what-are-the-chances/#:~:text=Union%20is%20denoted%20by%20the,intersection%20of%20the%20two%20sets.)).
