### 2020-06-29 #359 @arendsee
> @arendsee Perhaps that sentence on the README is not suited for some logics or types of fuzzy sets. I am struggling to introduce the topic for people not familiar with fuzzy sets and at the same time support and explain what the package can do to more expert users. I'll carefully review the wording around the subject.

The first sentence in the README does not apply to the **specific** logic you are using in the package. That operator you are using will not return membership probabilities. Using that operator to infer union membership probability will give you a meaningless result.

> About probabilities of A or B, it is still unclear to me over what do you calculate that probability or when would you like to do this.

I'm referring to the union operator as it is used in your fuzzy vignette:

``` R
set.seed(4567) # To be able to have exact replicates
relations <- data.frame(sets = c(rep("A", 5), "B", "C"),
                          elements = c(letters[seq_len(6)], letters[6]),
                          fuzzy = runif(7))
fuzzy_set <- tidySet(relations)
```

```
> fuzzy_set
  elements sets     fuzzy
1        a    A 0.2309186
2        b    A 0.7412554
3        c    A 0.1833834
4        d    A 0.4221682
5        e    A 0.5996399
6        f    B 0.2773313
7        f    C 0.7307312
> BaseSet::union(fuzzy_set, sets = c("C", "B"))
  elements sets     fuzzy
1        f  C∪B 0.7307312
```

The union function is intended to calculate the probability that f is in C or B.

`P(f in B) = 0.277`
`P(f in C) = 0.731`

Now, from the laws of probability: `P(f in (B or C)) = 0.277 + 0.731 - P(f in (B and C))`
If we assume independence (which is VERY dicey), then `P(f in (B or C)) = 0.277 + 0.731 - 0.277 * 0.731`

**Your union function calculates the wrong *probability*.**

If you want to use the current logic, then you need to strongly state that the fuzzy values are NOT probabilities.

> For @arendsee's tall OR strong example, he seems to be illustrating the different calculation and final value in:
>
> Case 1) if tall (0.3) and strong (0.6) are treated as fuzzy truth values:
> then tall OR strong = max (tall, strong) = max(0.3, 0.6) = 0.6
>
> Case 2) if tall (0.3) and strong (0.6) are treated as probabilities (and assume independence): **
> then tall OR strong = P(tall) + P(strong) - P(tall and strong) = 0.3 + 0.6 - (0.3 x 0.6) = 0.72

@j23414 I think I didn't explain my point here clearly. The tall or strong example was purely fuzzy. The data is fuzzy and probabilistic methods do not apply. I was describing two options for operators over fuzzy data. The first is the "standard" fuzzy operator that is currently used in BaseSet. The second is another fuzzy operator that happens to be the same as the probability of independent events, but the interpretation is different. Specifically, 0.72 is not the probability that the statement, "I am tall or strong", is true, rather it is a measure of how true the statement is. These two cases show two artificial metrics for fuzziness.

My point was that there are different fuzzy union operators. One of the fuzzy operators looks like the probabilistic operator for independent events. But just because this one operator can be used for set membership (assuming independence), does not mean *any* fuzzy operator can be used. There are many fuzzy metrics, but only one correct probability.

We don't get to choose whether data we are given is probabilities or fuzzy truth values.

> While users should not use different logics on the same data the package aims to allow them to use whatever logic they want according to the knowledge they have about the data. But I can't guide them or provide feedback.

The purpose of the README and vignettes is to guide them. But your examples, both in the README and vignettes, incorrectly use *fuzzy* operators to infer the *probabilities* of memberships in a set union.

> The fuzzy set theory extends the probability framework as explained on a source @j23414 provided:
>
> > ... fuzzy researchers have gone to great pains to distance
> > themselves from probability. But in so doing, many of them have lost track
> > of another point, which is that the converse DOES hold: all probability
> > distributions are fuzzy sets! As fuzzy sets and logic generalize Boolean
> > sets and logic, they also generalize probability.

Yes, fuzzy set theory is MORE general than probability. This means that probability is a special case of fuzzy logic. Thus a specific logic within fuzzy set theory applies to probabilistic systems. This does not mean that probability theory equals fuzzy theory. And it emphatically does not mean that any fuzzy logic applies to probability. Specifically, the fuzzy operator you are using for unions is not the same as the probability operator.

** in probability theory: ** `P(A or B) = P(A) + P(B) - P(A and B)`
** in standard fuzzy theory: ** ` `F(A or B) = max(F(A), F(B))`

P(A or B) is not equal to F(A or B), therefore you simply can't use that operator to calculate the probability of membership in a set union.

Maybe we should invite a statistician and/or a fuzzy set theorist to comment.
