+++ R/print.R
@@ -4,7 +4,8 @@ NULL
 #' Method to show the TidySet object
 #'
 #' Prints the resulting table of a TidySet object. Does not shown elements or
-#' sets without any relationship (empty sets). To see them use [sets()] or [elements()].
+#' sets without any relationship (empty sets). To see them use [sets()] or
+#' [elements()].
 #' @param object A TidySet
 #'
 #' @return A table with the information of the relationships.
