### 2020-06-30 d852f9f
Remove some long lines
Also removing some parenthesis to avoid creating a weird generic. Closes #48
