+++ R/AllGenerics.R
@@ -87,7 +87,9 @@ methods::setGeneric("nElements", function(object) standardGeneric("nElements"))
 #' @examples
 #' TS <- tidySet(list(A = letters[1:2], B = letters[5:7]))
 #' nRelations(TS)
-methods::setGeneric("nRelations", function(object) standardGeneric("nRelations"))
+methods::setGeneric("nRelations", function(object)
+  standardGeneric("nRelations")
+)
 
 #' Incidence
 #'
@@ -117,17 +119,19 @@ methods::setGeneric("relations", function(object) standardGeneric("relations"))
 
 #' @rdname relations
 #' @export
-methods::setGeneric("relations<-", function(object, value) {
+methods::setGeneric("relations<-", function(object, value)
       standardGeneric("relations<-")
-  })
+)
 
 #' @rdname elements
 #' @export
-methods::setGeneric("elements<-", function(object, value) standardGeneric("elements<-"))
+methods::setGeneric("elements<-", function(object, value)
+  standardGeneric("elements<-"))
 
 #' @rdname sets
 #' @export
-methods::setGeneric("sets<-", function(object, value) standardGeneric("sets<-"))
+methods::setGeneric("sets<-", function(object, value)
+  standardGeneric("sets<-"))
 
 #' Calculates the cardinality of a set
 #'
@@ -140,7 +144,8 @@ methods::setGeneric("sets<-", function(object, value) standardGeneric("sets<-"))
 #' @aliases cardinality
 #' @family sizes
 #' @family methods
-methods::setGeneric("set_size", function(object, set = NULL) standardGeneric("set_size"))
+methods::setGeneric("set_size", function(object, set = NULL)
+  standardGeneric("set_size"))
 
 #' Calculates the size of the elements
 #'
@@ -152,9 +157,9 @@ methods::setGeneric("set_size", function(object, set = NULL) standardGeneric("se
 #' @family sizes
 #' @export element_size
 #' @family methods
-methods::setGeneric("element_size", function(object, element = NULL) {
+methods::setGeneric("element_size", function(object, element = NULL)
       standardGeneric("element_size")
-  })
+)
 
 #' Rename sets
 #'
@@ -168,9 +173,9 @@ methods::setGeneric("element_size", function(object, element = NULL) {
 #' @seealso \code{\link{name_sets}}
 #' @export
 #' @family methods
-methods::setGeneric("rename_set", function(object, old, new) {
+methods::setGeneric("rename_set", function(object, old, new)
       standardGeneric("rename_set")
-  })
+)
 
 #' Rename elements
 #'
@@ -184,9 +189,9 @@ methods::setGeneric("rename_set", function(object, old, new) {
 #' @seealso \code{\link{name_elements}}
 #' @export
 #' @family methods
-methods::setGeneric("rename_elements", function(object, old, new) {
+methods::setGeneric("rename_elements", function(object, old, new)
       standardGeneric("rename_elements")
-  })
+)
 
 #' Name sets
 #'
@@ -205,9 +210,9 @@ methods::setGeneric("rename_elements", function(object, old, new) {
 #' )
 #' a <- tidySet(relations)
 #' name_sets(a)
-methods::setGeneric("name_sets", function(object) {
+methods::setGeneric("name_sets", function(object)
       standardGeneric("name_sets")
-  })
+)
 
 #' Name elements
 #'
@@ -225,9 +230,9 @@ methods::setGeneric("name_sets", function(object) {
 #' )
 #' a <- tidySet(relations)
 #' name_elements(a)
-methods::setGeneric("name_elements", function(object) {
+methods::setGeneric("name_elements", function(object)
[…]
