+++ man/rename_elements.Rd
@@ -22,9 +22,9 @@ A \code{TidySet} object.
 \description{
 Change the default names of sets and elements.
 }
-\section{Functions}{
+\section{Methods (by class)}{
 \itemize{
-\item \code{rename_elements,TidySet-method}: Rename elements
+\item \code{TidySet}: Rename elements
 }}
 
 \seealso{
