### 2020-06-30 27da142
Add argument for custom function for complement.
Related to #41 and #39
