### 2020-06-30 ee4189d
Make the examples to follow the standard
