+++ man/complement.Rd
@@ -21,24 +21,24 @@ activate with complement or use the specific function. You must specify if
 you want the complements of sets or elements.
 }
 \examples{
-relations <- data.frame(
-    sets = c("a", "a", "b", "b", "c", "c"),
+rel <- data.frame(
+    sets = c("A", "A", "B", "B", "C", "C"),
     elements = letters[seq_len(6)],
     fuzzy = runif(6)
 )
-a <- tidySet(relations)
-a \%>\%
+TS <- tidySet(rel)
+TS \%>\%
     activate("elements") \%>\%
     complement("a")
-a \%>\%
+TS \%>\%
     activate("elements") \%>\%
     complement("a", "C_a", keep = FALSE)
-a \%>\%
+TS \%>\%
     activate("set") \%>\%
-    complement("a")
-a \%>\%
+    complement("A")
+TS \%>\%
     activate("set") \%>\%
-    complement("a", keep = FALSE)
+    complement("A", keep = FALSE)
 }
 \seealso{
 \code{\link{activate}}
