### 2020-06-30 1e4b1cf
Add examples to droplevels.TidySet and test it
