### 2020-06-30 feab750
Set fuzzy to the length of the relations
To allow for empty relations
