### 2020-06-30 b3dd535
Better document what union, intersection and complement do.
Adding more examples.
