### 2020-06-30 923fb34
Fix? test, when input is factor
