### 2020-06-30 d4eafac
Simplify the assignment
