### 2020-06-30 5f88c5a
More detailed error messages when renaming
