### 2020-06-30 eb88cd8
Allowing empty lists
