### 2020-06-30 4976af7
Improve title of intersection and union
