+++ R/AllGenerics.R
@@ -323,10 +323,12 @@ methods::setGeneric("remove_column", function(object, slot, column_names)
       standardGeneric("remove_column")
 )
 
-#' Intersection of two sets
+#' Intersection of two or more sets
 #'
-#' Given a TidySet creates a new set with the elements on the both of them.
+#' Given a TidySet creates a new set with the elements on the both of them
+#' following the logic defined on FUN.
 #' @param sets The character of sets to be intersect.
+#' @param name The name of the new set. By defaults joins the sets with an "∩".
 #' @inheritParams union
 #' @param FUN A function to be applied when performing the union.
 #' The standard intersection is the "min" function, but you can provide any
@@ -338,13 +340,22 @@ methods::setGeneric("remove_column", function(object, slot, column_names)
 #' @family methods that create new sets
 #' @family methods
 #' @examples
-#' relations <- data.frame(
-#'     sets = c(rep("a", 5), "b"),
+#' rel <- data.frame(
+#'     sets = c(rep("A", 5), "B"),
+#'     elements = c("a", "b", "c", "d", "f", "f")
+#' )
+#' TS <- tidySet(rel)
+#' intersection(TS, c("A", "B")) # Default Name
+#' intersection(TS, c("A", "B"), "C") # Set the name
+#' # Fuzzy set
+#' rel <- data.frame(
+#'     sets = c(rep("A", 5), "B"),
 #'     elements = c("a", "b", "c", "d", "f", "f"),
 #'     fuzzy = runif(6)
 #' )
-#' a <- tidySet(relations)
-#' intersection(a, c("a", "b"), "C")
+#' TS2 <- tidySet(rel)
+#' intersection(TS2, c("A", "B"), "C")
+#' intersection(TS2, c("A", "B"), "C", FUN = function(x){max(sqrt(x))})
 methods::setGeneric("intersection", function(object, sets, ...)
       standardGeneric("intersection")
 )
