### 2020-06-30 1066673
Remove some checks to allow empty TidySets
