### 2020-06-30 66cfd03
Change test to recent removed checks
