### 2020-06-30 bb91dae
Fix test for more restrictive naming
