+++ tests/testthat/test-name.R
@@ -23,7 +23,8 @@ test_that("name_sets<- works", {
     out <- name_sets(a)
     expect_equal(out, LETTERS[1:2])
     b <- add_column(a, "sets", data.frame(Ha = c("a", "b")))
-    expect_error(name_sets(b) <- "A2", "Duplicated sets")
+    expect_error(name_sets(b) <- c("A2", "A2"), "Duplicated sets")
+    expect_error(name_sets(b) <- "A2", "Less names")
     expect_error(name_sets(a) <- as.factor(letters[1:2]), NA)
 })
 
@@ -52,5 +53,7 @@ test_that("name_elements<- works", {
     expect_equal(out, LETTERS[1:6])
 
     b <- add_column(a, "elements", data.frame(Ha = paste0(letters[1:6], "2")))
-    expect_error(name_elements(b) <- "A2", "Duplicated elements")
+    expect_error(name_elements(b) <- rep("A2", nElements(b)),
+                 "Duplicated elements")
+    expect_error(name_elements(b) <- "A2", "Less names")
 })
