### 2020-06-30 6818164
Add cardinality method
