### 2020-06-30 cd5588a
Adding examples and slight touches on descriptions.
