+++ R/GeneSetCollection.R
@@ -12,6 +12,20 @@ tidy <- function(object) {
 #' @describeIn tidy Converts to a tidySet given a GeneSetCollection
 #' @export
 #' @method tidy GeneSetCollection
+#' @examples
+#' if (require(GSEABase)) {
+#'     library("GSEABase")
+#' } else {
+#'     stop("For this example you need GSEABase package from Bioconductor.")
+#' }
+#' gs <- GeneSet()
+#' gs
+#' tidy(gs)
+#' fl <- system.file("extdata", "Broad.xml", package="GSEABase")
+#' gs2 <- getBroadSets(fl) # actually, a list of two gene sets
+#' TS <- tidy(gs2)
+#' dim(TS)
+#' sets(TS)
 tidy.GeneSetCollection <- function(object) {
     data <- slot(object, ".Data")
     sets <- lapply(data, tidy)
