+++ R/size.R
@@ -1,24 +1,36 @@
 #' Size
 #'
-#' Calculate the size of the elements or sets. First it must have active either
-#' sets or elements.
+#' Calculate the size of the elements or sets, using the fuzzy values as
+#' probabilities. First it must have active either sets or elements.
 #' @param object A TidySet object
 #' @param ... Character vector with the name of elements or sets you want to
 #' calculate the size of.
 #' @export
+#' @seealso A related concept [cardinality()]. It is calculated using
+#' [length_set()].
 #' @return The size of the elements or sets. If there is no active slot or it
 #' is the relations slot returns the TidySet object with a warning.
 #' @examples
+#' rel <- data.frame(
+#'     sets = c(rep("A", 5), "B", "C"),
+#'     elements = c(letters[seq_len(6)], letters[6])
+#' )
+#' TS <- tidySet(rel)
+#' TS <- activate(TS, "elements")
+#' size(TS)
+#' TS <- activate(TS, "sets")
+#' size(TS)
+#' # With fuzzy sets
 #' relations <- data.frame(
 #'     sets = c(rep("A", 5), "B", "C"),
 #'     elements = c(letters[seq_len(6)], letters[6]),
 #'     fuzzy = runif(7)
 #' )
-#' a <- tidySet(relations)
-#' a <- activate(a, "elements")
-#' size(a)
-#' a <- activate(a, "sets")
-#' size(a)
+#' TS <- tidySet(relations)
+#' TS <- activate(TS, "elements")
+#' size(TS)
+#' TS <- activate(TS, "sets")
+#' size(TS)
 size <- function(object, ...) {
     UseMethod("size")
 }
