+++ man/union.Rd
@@ -47,6 +47,10 @@ A \code{TidySet} object.
 Given a TidySet merges several sets into the new one using the logic
 defined on FUN.
 }
+\details{
+The default uses the \code{max} function following the \href{https://doi.org/10.1016/S0019-9958(65)90241-X}{standard fuzzy definition}, but it can be
+changed.
+}
 \examples{
 # Classical set
 rel <- data.frame(
