+++ man/intersection.Rd
@@ -47,6 +47,10 @@ A \code{TidySet} object.
 Given a TidySet creates a new set with the elements on the both of them
 following the logic defined on FUN.
 }
+\details{
+#' The default uses the \code{min} function following the \href{https://doi.org/10.1016/S0019-9958(65)90241-X}{standard fuzzy definition}, but it can be
+changed.
+}
 \section{Methods (by class)}{
 \itemize{
 \item \code{object = TidySet,sets = character}: Applies the standard intersection
