### 2020-06-30 daa56c9
Update vignettes with more explanations
