### 2020-07-06 ba42a25
Easier to find intersection
