### 2020-07-06 1b88fc2
Adding more links and commenting more the results of the vignettes
