+++ vignettes/fuzzy.Rmd
@@ -49,14 +49,17 @@ When does the switch happen from warm to hot?
 In fuzzy set theories the step from a continuous scale to a categorical scale is performed by the [membership function](https://en.wikipedia.org/wiki/Membership_function_(mathematics)) and is called [fuzzification](https://en.wikipedia.org/wiki/Fuzzy_logic#Fuzzification). 
 
 When there is a degree of membership and uncertainty on the membership it is considered a [type-2 fuzzy set](https://en.wikipedia.org/wiki/Type-2_fuzzy_sets_and_systems).
-We can understand it with using as example a paddel ball, it can be used for tennis or paddel (membership), but until we don't test it bouncing on the court we won't be sure (assignment) if it is a paddel ball or a tennis ball. 
-We could think about a ball equally good for tennis and paddel (membership) but which two people thought it is for tennis and the other for paddel. 
+We can understand it with using as example a paddle ball, it can be used for tennis or paddle (membership), but until we don't test it bouncing on the court we won't be sure (assignment) if it is a paddle ball or a tennis ball. 
+We could think about a ball equally good for tennis and paddle (membership) but which two people thought it is for tennis and the other for paddle. 
 
 These voting/rating system is also a common scenario where fuzzy sets arise.
 When several graders/people need to agree but compromise on a middle ground. 
+One modern example of this is ratting apps where several people vote between 1 and 5 an app and the displayed value takes into consideration all the votes. 
+
+As you can see when one does have some vagueness or uncertainty then fuzzy logic is a good choice.
+There have been developed several logic and methods. 
 
 # Creating a fuzzy set
-
 
 To create a fuzzy set you need to have a column named "fuzzy" if you create it 
 from a `data.frame` or have a named numeric vector if you create it from a `list`.
@@ -82,22 +85,24 @@ We can work with fuzzy sets as we do with normal sets. But if you remember that 
 You can make a union of two sets present on the same object.
 
 ```{r union}
-BaseSet::union(fuzzy_set, sets = c("C", "B"), name = "D")
-BaseSet::union(fuzzy_set, sets = c("C", "B"), name =  "D", FUN = "mean")
+BaseSet::union(fuzzy_set, sets = c("A", "B"))
+BaseSet::union(fuzzy_set, sets = c("A", "B"), name = "D")
 ```
 
-Observe the fuzziness on the resulting set has changed (on the traditional sets it doesn't matter)
+We get a new set with all the elements on both sets.
+There isn't an element in both sets A and B, so the fuzzy values here do not change.
 
 ## Intersection
 
-We can do the same with the intersection
+However if we do the intersection between B and C we can see some changes on the fuzzy value:
 
 ```{r intersection}
-intersection(fuzzy_set, sets = c("A", "B"), keep = FALSE)
-intersection(fuzzy_set, sets = c("A", "B"), keep = FALSE, FUN = "mean")
+intersection(fuzzy_set, sets = c("B", "C"), keep = FALSE)
+intersection(fuzzy_set, sets = c("B", "C"), keep = FALSE, FUN = "mean")
 ```
 
-But in this example we don't see any difference on the calculated values.
+Different logic on the `union()`, `intersection()`, `complement()`, and `cardinality()`, we get different fuzzy values. 
+Depending on the nature of our fuzziness and the intended goal we might apply one or other rules.
 
 ## Complement
 
@@ -107,10 +112,11 @@ We can look for the complement of one or several sets:
 complement_set(fuzzy_set, sets = "A", keep = FALSE)
 ```
 
-Note that the values of the complement are `1-fuzzy`:
+Note that the values of the complement are `1-fuzzy` but can be changed:
 
 ```{r complement_previous}
-fuzzy_set %>% filter(sets == "A")
+filter(fuzzy_set, sets == "A")
[…]
