+++ vignettes/advanced.Rmd
@@ -41,9 +41,11 @@ library("dplyr")
 library("BaseSet")
 ```
 
-# Human GO
+# Human gene ontology
 
-We will explore the genes related with GO. 
+We will explore the genes with assigned gene ontology terms. 
+These terms describe what is the process and role of the genes.
+The links are annotated with different [evidence codes](http://geneontology.org/docs/guide-go-evidence-codes/) to indicate how such annotation is supported. 
 
 ```{r prepare_GO, message=FALSE}
 # We load some libraries
@@ -54,24 +56,27 @@ library("ggplot2")
 h2GO_TS <- tidySet(org.Hs.egGO)
 h2GO <- as.data.frame(org.Hs.egGO)
 ```
-We can now explore if there are differences in evidence usage for each ontology in GO:
+
+We can now explore if there are differences in evidence usage for each ontology in gene ontology:
+
 ```{r evidence_ontology}
 library("forcats")
 h2GO %>% 
-    group_by(Evidence, Ontology) %>% 
-    count(name = "Freq") %>% 
-    ungroup() %>% 
-    mutate(Evidence = fct_reorder2(Evidence, Ontology, -Freq),
-           Ontology = case_when(Ontology == "CC" ~ "Cellular Component",
-                                Ontology == "MF" ~ "Molecular Function",
-                                Ontology == "BP" ~ "Biological Process",
-                                TRUE ~ NA_character_)) %>% 
-    ggplot() +
-    geom_col(aes(Evidence, Freq)) +
-    facet_grid(~Ontology) + 
-    theme_bw() +
-    coord_flip() +
-    xlab("Evidence")
+  group_by(Evidence, Ontology) %>% 
+  count(name = "Freq") %>% 
+  ungroup() %>% 
+  mutate(Evidence = fct_reorder2(Evidence, Ontology, -Freq),
+         Ontology = case_when(Ontology == "CC" ~ "Cellular Component",
+                              Ontology == "MF" ~ "Molecular Function",
+                              Ontology == "BP" ~ "Biological Process",
+                              TRUE ~ NA_character_)) %>% 
+  ggplot() +
+  geom_col(aes(Evidence, Freq)) +
+  facet_grid(~Ontology) + 
+  theme_minimal() +
+  coord_flip() +
+  labs(x = element_blank(), y = element_blank(),
+       title = "Evidence codes for each ontology")
 ```
 
 We can see that biological process are more likely to be defined by IMP evidence code that means inferred from mutant phenotype. 
@@ -83,14 +88,16 @@ This graph doesn't consider that some relationships are better annotated than ot
 h2GO_TS %>% 
   relations() %>% 
   group_by(elements, sets) %>% 
-  count(sort = TRUE) %>% 
+  count(sort = TRUE, name = "Annotations") %>% 
+  ungroup() %>% 
+  count(Annotations, sort = TRUE) %>% 
   ggplot() +
-  geom_histogram(aes(n), binwidth = 1) +
+  geom_col(aes(Annotations, n)) +
   theme_minimal() +
-  # scale_y_log10() +
-  labs(x = "Evidence codes", 
-       title = "Evidence codes for each relation between GO and a gene",
-       subtitle = "in human annotation")
+  labs(x = "Evidence codes", y = "Annotations", 
+       title = "Evidence codes for each annotation",
+       subtitle = "in human") +
+  scale_x_continuous(breaks = 1:7)
 ```
 
 We can see that mostly all the annotations are done with a single evidence code.
@@ -113,7 +120,7 @@ sum(sizes_set$size == 0)
[…]
