### 2020-07-06 bd3615d
Add why to use the package
