### 2020-07-06 bd52da1
Add reviewers (also fix github actions pandoc installation)
