+++ .github/workflows/R-CMD-check.yaml
@@ -12,7 +12,7 @@ jobs:
         os: [macos-latest, windows-latest, ubuntu-latest]
     steps:
       - uses: actions/checkout@vmaster
-      - uses: r-lib/actions/setup-pandoc@v1
+      - uses: r-lib/actions/setup-pandoc@master
       with:
         pandoc-version: '2.10' # The pandoc version to download
       - uses: r-lib/actions/setup-r@master
