### 2020-07-06 c3e6ef3
Make sure that yaml is valid
