+++ .github/workflows/R-CMD-check.yaml
@@ -14,10 +14,11 @@ jobs:
       - uses: actions/checkout@master
       - uses: r-lib/actions/setup-pandoc@master
       - name: "Install pandoc"
-      with:
-        pandoc-version: '2.10' # The pandoc version to download
+        with:
+          pandoc-version: '2.10' # The pandoc version to download
       - run: echo "# Test" | pandoc -t html
       - uses: r-lib/actions/setup-r@master
+      - name: "Setting up R"
       - name: Install Linux dependencies
         if: runner.os == 'Linux'
         run: sudo apt install libcurl4-openssl-dev
