+++ .github/workflows/R-CMD-check.yaml
@@ -18,7 +18,6 @@ jobs:
           pandoc-version: '2.10' # The pandoc version to download
       - run: echo "# Test" | pandoc -t html
       - uses: r-lib/actions/setup-r@master
-      - name: "Setting up R"
       - name: Install Linux dependencies
         if: runner.os == 'Linux'
         run: sudo apt install libcurl4-openssl-dev
