+++ .github/workflows/R-CMD-check.yaml
@@ -11,14 +11,14 @@ jobs:
       matrix:
         os: [macos-latest, windows-latest, ubuntu-latest]
     steps:
-      - uses: actions/checkout@master
-      - uses: r-lib/actions/setup-pandoc@master
+      - uses: actions/checkout@v1
       - name: "Install pandoc"
+        uses: r-lib/actions/setup-pandoc@master
         with:
           pandoc-version: '2.10' # The pandoc version to download
-      - run: echo "# Test" | pandoc -t html
-      - uses: r-lib/actions/setup-r@master
+        run: echo "# Test" | pandoc -t html
       - name: Install Linux dependencies
+        uses: r-lib/actions/setup-r@master
         if: runner.os == 'Linux'
         run: sudo apt install libcurl4-openssl-dev
       - name: Install basic dependencies
