+++ .github/workflows/check-full.yaml
@@ -6,7 +6,7 @@ on:
     branches:
       - master
 
-name: R-CMD-check
+name: R-CMD-check-full
 
 jobs:
   R-CMD-check:
