+++ .github/workflows/R-CMD-check.yaml
@@ -15,14 +15,13 @@ jobs:
       - uses: r-lib/actions/setup-pandoc@master
         with:
           pandoc-version: '2.10' # The pandoc version to download
-        run: echo "# Test" | pandoc -t html
+      - run: echo "# Test" | pandoc -t html
       - name: Install Linux dependencies
-        uses: r-lib/actions/setup-r@master
+      - uses: r-lib/actions/setup-r@master
         if: runner.os == 'Linux'
-        run: sudo apt install libcurl4-openssl-dev
+      - run: sudo apt install libcurl4-openssl-dev
       - name: Install basic dependencies
         run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'BiocManager'))"
-  check:
       - name: Install all dependencies
         run: Rscript -e "BiocManager::install('llrs/BaseSet', dependencies = TRUE, build_vignettes = TRUE)"
       - name: Session Info
