+++ .github/workflows/R-CMD-check.yaml
@@ -22,11 +22,10 @@ jobs:
       - uses: r-lib/actions/setup-pandoc@master
         with:
           pandoc-version: '2.10' # The pandoc version to download
-      - run: echo "# Test" | pandoc -t html
       - name: Install Linux dependencies
         uses: r-lib/actions/setup-r@master
         if: runner.os == 'Linux'
-      - run: sudo apt install libcurl4-openssl-dev
+        run: sudo apt install libcurl4-openssl-dev
       - name: Install basic dependencies
         run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'BiocManager'))"
       - name: Install all dependencies
