### 2020-07-07 45b513e
Update code/vignette
