+++ man/BaseSet-package.Rd
@@ -23,6 +23,12 @@ Useful links:
 
 }
 \author{
-\strong{Maintainer}: Lluís  Revilla Sancho \email{lluis.revilla@gmail.com}
+\strong{Maintainer}: Lluís Revilla Sancho \email{lluis.revilla@gmail.com} (\href{https://orcid.org/0000-0001-9747-2570}{ORCID})
+
+Other contributors:
+\itemize{
+  \item Zebulun Arendsee [reviewer]
+  \item Jennifer Chang [reviewer]
+}
 
 }
