### 2020-07-08 93baa85
Protection for when multiple sizes are calculated
