### 2020-07-08 e351ca6
Change examples to follow standard notation (upper case sets, lower case elements)
Also added links to how are *_size calculated as well as changing from singular to plural the argument
