+++ R/AllGenerics.R
@@ -48,6 +48,9 @@ methods::setGeneric("sets", function(object) standardGeneric("sets"))
 #' @return A logical value.
 #' @export is.fuzzy
 #' @family methods
+#' @examples
+#' TS <- tidySet(list(A = letters[1:5], B = letters[2:10]))
+#' is.fuzzy(TS)
 methods::setGeneric("is.fuzzy", function(object) standardGeneric("is.fuzzy"))
 
 #' Number of sets
@@ -137,26 +140,26 @@ methods::setGeneric("sets<-", function(object, value)
 #'
 #' Calculate the probability of being of different sizes for a given set.
 #' @param object A TidySet object.
-#' @param set The sets from which the length is calculated.
+#' @param sets The sets from which the length is calculated.
 #' @return A list with the size of the set or the probability of having that
 #' size.
 #' @export set_size
 #' @family sizes
 #' @family methods
-methods::setGeneric("set_size", function(object, set = NULL)
+methods::setGeneric("set_size", function(object, sets = NULL)
   standardGeneric("set_size"))
 
 #' Calculates the size of the elements
 #'
 #' Calculate the probability of being of different sizes for a given set.
 #' @param object A TidySet object.
-#' @param element The element from which the length is calculated.
+#' @param elements The element from which the length is calculated.
 #' @return A list with the size of the elements or the probability of having
 #' that size.
 #' @family sizes
 #' @export element_size
 #' @family methods
-methods::setGeneric("element_size", function(object, element = NULL)
+methods::setGeneric("element_size", function(object, elements = NULL)
       standardGeneric("element_size")
 )
 
@@ -215,12 +218,12 @@ methods::setGeneric("rename_elements", function(object, old, new)
 #' @export
 #' @examples
 #' relations <- data.frame(
-#'     sets = c(rep("a", 5), "b"),
+#'     sets = c(rep("A", 5), "B"),
 #'     elements = letters[seq_len(6)],
 #'     fuzzy = runif(6)
 #' )
-#' a <- tidySet(relations)
-#' name_sets(a)
+#' TS <- tidySet(relations)
+#' name_sets(TS)
 methods::setGeneric("name_sets", function(object)
       standardGeneric("name_sets")
 )
@@ -235,12 +238,12 @@ methods::setGeneric("name_sets", function(object)
 #' @export
 #' @examples
 #' relations <- data.frame(
-#'     sets = c(rep("a", 5), "b"),
+#'     sets = c(rep("A", 5), "B"),
 #'     elements = letters[seq_len(6)],
 #'     fuzzy = runif(6)
 #' )
-#' a <- tidySet(relations)
-#' name_elements(a)
+#' TS <- tidySet(relations)
+#' name_elements(TS)
 methods::setGeneric("name_elements", function(object)
       standardGeneric("name_elements")
 )
@@ -258,13 +261,13 @@ methods::setGeneric("name_elements", function(object)
 #' @aliases name_elements<-
 #' @examples
 #' relations <- data.frame(
-#'     sets = c(rep("a", 5), "b"),
+#'     sets = c(rep("A", 5), "B"),
 #'     elements = letters[seq_len(6)],
 #'     fuzzy = runif(6)
 #' )
-#' a <- tidySet(relations)
-#' a
-#' name_elements(a) <- letters[1:6]
+#' TS <- tidySet(relations)
+#' TS
+#' name_elements(TS) <- letters[1:6]
[…]
