+++ man/name_elements.Rd
@@ -32,12 +32,12 @@ Retrieve the name of the elements.
 
 \examples{
 relations <- data.frame(
-    sets = c(rep("a", 5), "b"),
+    sets = c(rep("A", 5), "B"),
     elements = letters[seq_len(6)],
     fuzzy = runif(6)
 )
-a <- tidySet(relations)
-name_elements(a)
+TS <- tidySet(relations)
+name_elements(TS)
 }
 \seealso{
 Other names: 
