### 2020-07-09 5a5ccb6
Further clarification of the union logic
