### 2020-07-10 064f68e
Remove conditonal, not sure if needed, because otherwise it would delte data
