### 2020-07-13 #359 @llrs
Review 2:

I'll only reply to comments or points that weren't checked or there were some substantial comments:

 - [x] Vignette(s) demonstrating major functionality that runs successfully locally
  Now the vignettes run succesfully at least on windows, ubuntu, and I think the problems on macOs of the CI are not related to the vignette.
 - [x] Examples for all exported functions in R Help that run successfully locally
  Corrected examples
 - [x] Automated tests: Unit tests cover essential functions of the package
  Fixed the failing tests and added some more
 - [x] Functionality: Any functional claims of the software been confirmed.

 - Pluralized arguments of `set_size` and `element_size` to match the expected name of columns
 - Modified the examples so that sets are represented by upper case letters and elements by lower case letters.
 - About the weird behavior of `set_size()`
Your first try to modify the set and convert to a fuzzy set you'll need to assign the modified ecoli_set to be able to use the fuzzy values (or pipe it as shown below). The TidySet is of class S4, not S6, and the modifications are not in place.

```r
ecoli_sets %>%
      mutate(
      fuzzy = case_when(sets == "GL" ~ 0.2,    # Add fuzzy values
      sets == "CF" ~ 0.8)) %>%
      set_size()
```

Yes, the GL has 0.04 probability to have two genes so the glycolysis pathway probably don't have any genes. That's is if we assume that the fuzzy values are probabilities, if they where other things it wouldn't be accurate. I think this is different from cardinality, as cardinality it is just a number for a set, while here I return several values for a single set. I added to the documentation how is this calculated (via `length_set`).

I provided a new method `cardinality` to calculate it and allow to provide a fuzzy logic function. The `{sets}` package mentioned previously provides cardinality as the number of elements in a set regardless of the fuzzy value. However, in this case the default is `sum` as you found, but `length` as in `{sets}` or other functions could be provided.

 - [x] Vignette comments

Changed the automatic names for the full names, also added some links for references to the data and how to interpret it. I tried to use some fuzzy values on the advanced vignette, but it would take too long to calculate for many sets and would make the package fail the checks.

----

**Fuzzy logic & explanation**

Many thanks for all the feedback on this, I realized that I assumed too many things. I modified or added explanations about the fuzzy values around the package.

I modified the README section about fuzzy sets, now it reads : "Fuzzy sets are similar to classical sets but there is some vagueness on the relationship between the element and the set.".

Similarly I've added a new section on the README about why this package was developed and why it could be useful to others. Also introduced the topic on the fuzzy vignette about what are they and some links.

I do not want to force an interpretation of the fuzzy theory which sometimes translate to using different semantics: membership, probabilities, truth vales, and others. I think that the fuzzy column is the most neutral and better understood. There seems to be a movement from fuzzy research to [differentiate it from probabilities](http://www.cs.cmu.edu/Groups/AI/html/faqs/ai/fuzzy/part1/faq-doc-10.html), but at the same time to [expand them](https://en.wikipedia.org/wiki/Fuzzy_logic#Comparison_to_probability)(probabilities).

One can use any logic according to the specific framework one wish to work. This in encouraged through the package by being able to use any logic. To show the flexibility of the package using the same data I've added some examples with a different logic explaining when this logic could be applied. Hope now  it is clear that the package does not forces to use a certain logic.

The defaults parameters of the package are meant to provide sensible results with fuzzy sets but to not give surprises on classical sets. On the specific case of union I added an example on the documentation that I hope it makes it clear how to use other logics outside the default for the specific case of probabilities, `union(TS, sets = c("A", "B"), FUN = function(x){y <- sum(x)-prod(x); ifelse(length(x)==1, x, y)})`.

Given all the discussion around how to interpret the fuzzy sets, I would appreciate your thoughts on when do fuzzy sets arise on practical cases. How do imagine a case where there are multiple cases which might be related or not to a set? For instance imagine I want to record some cards on an intersection; the first blue car goes right, second blue car goes right, third blue car goes left. Would you record as a fuzzy/probability set (set being here right/left) where a blue car has 0.66 relationship with going right and 0.33 going left? Or as a classical set with all three cars two of them on the right set and one on the left set? If it is a fuzzy set, then which logic would you apply?
