### 2020-07-13 #359 @arendsee
The term `sum(x) - prod(x)` is only equal to the probability of the union when x has length 2. Calculating the union probabilities is more complicated for more than 2 sets. For 3 sets it is `P(A or B or C) = P(A) + P(B) + P(C) - P(A and B) - P(B and C) - P(C and A) + P(A and B and C)`. I recommend adding test cases for sets of 2, 3 and 4 (at least).

Think about the case of three sets with probability of membership for each being 1. The probability of the union would also be 1. But `sum(x) - prod(x) == 3 - 1 == 2`, which of course isn't a valid probability. In general, the union of any vector of numbers between 0 and 1 should also be between 0 and 1 (you could write this into a unit test).

I don't think it is necessary for initial acceptance into rOpenSci, but you should eventually consider how to relax the independence assumption. I'd definitely recommend talking to a statistician (statisticians are great).

I think fuzzy sets have lots of applications within bioinformatics. For example, someone might want to know whether metabolite A and metabolite B both have high concentrations in a given tissue. You could scale the molar concentrations to values between 0 and 1 where 0.5 is the "normal" value.
