### 2020-07-14 6fb5789
Add helpers for union using indepdenent probabilities
