+++ man/length_probability.Rd
@@ -1,9 +1,12 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/length.R
-\name{length_probability}
+\name{union_probability}
+\alias{union_probability}
 \alias{length_probability}
 \title{Calculates the probability of a single length}
 \usage{
+union_probability(p)
+
 length_probability(p, size)
 }
 \arguments{
@@ -17,7 +20,13 @@ A numeric value of the probability of the given size.
 \description{
 Creates all the possibilities and then add them up.
 }
+\section{Functions}{
+\itemize{
+\item \code{union_probability}: Probability of independent values
+}}
+
 \examples{
+union_probability(c(0.5, 0.1, 0.3, 0.5, 0.25, 0.23))
 length_probability(c(0.5, 0.75), 2)
 length_probability(c(0.5, 0.75, 0.66), 1)
 length_probability(c(0.5, 0.1, 0.3, 0.5, 0.25, 0.23), 2)
