+++ NAMESPACE
@@ -91,6 +91,7 @@ export(group)
 export(group_by)
 export(incidence)
 export(independent)
+export(independent_probabilities)
 export(intersection)
 export(is.fuzzy)
 export(is_nested)
@@ -134,6 +135,7 @@ export(subtract)
 export(tidy)
 export(tidySet)
 export(union)
+export(union_probability)
 exportClasses(TidySet)
 exportMethods("elements<-")
 exportMethods("relations<-")
