### 2020-07-14 e806cc6
Improve documentation for union_probability
