### 2020-07-14 bc1f608
Slightly speed improvments by passing the function to combn
