### 2020-07-14 09e9609
Document what each function does
