+++ man/multiply_probabilities.Rd
@@ -19,12 +19,8 @@ A numeric value of the probability.
 }
 \description{
 Calculates the probability that all happened simultaneously
+\code{independent_probabilities} just multiply the probabilities of the index passed.
 }
-\section{Functions}{
-\itemize{
-\item \code{independent_probabilities}: Product of the probabilities selected.
-}}
-
 \examples{
 multiply_probabilities(c(0.5, 0.1, 0.3, 0.5, 0.25, 0.23), c(1, 3))
 independent_probabilities(c(0.5, 0.1, 0.3, 0.5, 0.25, 0.23), c(1, 3))
