### 2020-07-14 #359 @llrs
@arendsee I added a helper function `union_probability` to calculate the union probabilities for an unlimited size (only tested for 3 and 4 though). However I feel that this is out of the scope of the package.

The validation of the TidySet class includes a test that check that fuzzy values are between 0 and 1 (both included). There is no need to add a unit test as it is automatically checked with each exported method (via `validObject`).

As previously said I don't assume independence or dependence of sets in the package, the package doesn't even assume that the fuzzy sets are probabilities! This is and will be the responsibility of the user, mainly because from the data stored the package can't make any educated guesses about the type of the relationships between sets. I'll ask a statistician about relaxing independence between sets.

I agree that they have lots of applications within bioinformatics, that's why I wrote the package: to make it easier to use them. Thank you for all the feedback to make it even easier to use it.

Could you answer the questions of the last paragraph on my previous comment?
