### 2020-07-28 #359 @llrs
Yes, along the review I made several questions which I didn't get an answer yet.

Also I don't know what was decided about reaching a statistician (from above):
> Hello all. It looks like some guidance from a statistician might indeed be a good idea. This is technically not part of package review but I'll see what other editors think and let you know what we can do.
