### 2020-07-29 #359 @arendsee
OK, the math checks out. Also, the extended vignette clarifies the fuzziness relationship.

I think this package will have a good future. There are a lot of neat directions you could take it. I think it will be good to share it with the community and get feedback. The package may incite a lot of debate, which is good. Once you and the community have hammered out the wrinkles and chosen the features and interface that is "best", maybe you can roll out a non-backwards compatible BaseSet2 package.

So, my final thoughts: the package is useful as is and will get much better as an active community grows around it.
