### 2020-08-03 #359 @j23414
The new version passes checks so far... took a while to build the advanced vignette but I can see that fuzzy values were added (good). Glad to see more explanations and links in README, Fuzzy, and Advanced. For `set_size`, the new `cardinality` function matches what I'd expect in fuzzy sets.

In summary, can see how the package may be useful. Agree that it can continue to be refined with an active community.
