### 2020-08-04 #359 @annakrystalli
Quick update from me too, I'm finding it hard to find someone that's either available (lots of folks are on holiday) or comfortable commenting on the issue.

In truth, statistical packages are technically out of scope for standard rOpenSci review. I think we felt at submission that set operations we pretty well established so I didn't anticipate this is as an issue. That's my bad but it's why we so far have avoided packages with statistical elements that might cause this sort of confusion.

Now, moving forward, I think the ideal situation for all would if we all came to a satisfactory agreement here. In particular, we always want reviews to be a positive experience and for that I feel it's important that the authors and reviewers are on the same page at the end of it. So I look forward to @arendsee's response and hope the further clarifications by @llrs have helped reach an agreement on the outstanding issues.

If however there is still disagreement (which is fine btw) what I suggest is that we ask for help in the rOpenSci statistical software review slack channel (the new under development initiative to expand review to statistical packages).

How does that sound for a plan?
