### 2020-08-04 #359 @annakrystalli
Approved! Thanks @llrs for submitting and @arendsee & @j23414 for your reviews! 🎉 🚀

To-dos:
    * replace your package docs URL with `https://docs.ropensci.org/package_name`
