### 2020-08-04 #359 @llrs
Many thanks @annakrystalli for approving the package. I already transferred the package and hope to finish all the other steps either today or tomorrow.

Soon I'll begin my holidays and won't be able to contribute with a post for the blog, but I'd like to write one after September @stefaniebutland. But I'm also hoping to present on the Rstudio conference on January... not sure when I'll manage to write it up.
