### 2020-08-04 f0fd704
Change url and template of the docs
Hope it works
