+++ README.md
@@ -4,17 +4,17 @@
 <!-- badges: start -->
 
 [![Travis build
-status](https://travis-ci.org/llrs/BaseSet.svg?branch=master)](https://travis-ci.org/llrs/BaseSet)
+status](https://travis-ci.org/ropensci/BaseSet.svg?branch=master)](https://travis-ci.org/ropensci/BaseSet)
 [![Actions
-CI](https://github.com/llrs/BaseSet/workflows/R-CMD-check/badge.svg)](https://github.com/llrs/BaseSet/actions?query=workflow%3AR-CMD-check)
+CI](https://github.com/ropensci/BaseSet/workflows/R-CMD-check/badge.svg)](https://github.com/ropensci/BaseSet/actions?query=workflow%3AR-CMD-check)
 [![Coverage
-status](https://codecov.io/gh/llrs/BaseSet/branch/master/graph/badge.svg)](https://codecov.io/github/llrs/BaseSet?branch=master)
+status](https://codecov.io/gh/ropensci/BaseSet/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/BaseSet?branch=master)
 [![Lifecycle:
 maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![Project Status: Active – The project has reached a stable, usable
 state and is being actively
 developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![](https://badges.ropensci.org/359_status.svg)](https://github.com/ropensci/software-review/issues/359)
+[![rOpenSci](https://badges.ropensci.org/359_status.svg)](https://github.com/ropensci/software-review/issues/359)
 <!-- badges: end -->
 
 # BaseSet
@@ -44,8 +44,8 @@ numeric data associated with sets…
 Watch BaseSet working on the [examples](#Examples) below and in the
 vignettes. You can also find [related packages](#Related-packages) and
 the differences with BaseSet. If you have some questions or bugs [open
-an issue](https://github.com/llrs/BaseSet/issues) (remember the [Code of
-Conduct](#Code-of-Conduct))
+an issue](https://github.com/ropensci/BaseSet/issues) (remember the
+[Code of Conduct](#Code-of-Conduct))
 
 # Installation
 
@@ -60,7 +60,7 @@ if (!require("BiocManager")) {
 ```
 
 You can install the latest version of BaseSet from
-[Github](https://github.com/llrs/BaseSet) with:
+[Github](https://github.com/ropensci/BaseSet) with:
 
 ``` r
 BiocManager::install("llrs/BaseSet", 
@@ -79,11 +79,11 @@ sets_analysis <- tidySet(sets)
 sets_analysis
 #>   elements sets fuzzy
 #> 1        a    A     1
-#> 2        a    B     1
-#> 3        b    A     1
-#> 4        c    A     1
-#> 5        d    A     1
-#> 6        e    A     1
+#> 2        b    A     1
+#> 3        c    A     1
+#> 4        d    A     1
+#> 5        e    A     1
+#> 6        a    B     1
 #> 7        f    B     1
 ```
 
@@ -116,13 +116,13 @@ intersection(sets_analysis, sets = c("A", "B"))
 intersection(sets_analysis, sets = c("A", "B"), name = "D", keep = TRUE) 
 #>   elements sets fuzzy
 #> 1        a    A     1
-#> 2        a    B     1
-#> 3        a    D     1
-#> 4        b    A     1
-#> 5        c    A     1
-#> 6        d    A     1
-#> 7        e    A     1
-#> 8        f    B     1
+#> 2        b    A     1
[…]
