### 2020-08-05 #359 @stefaniebutland
> I'd like to write one after September @stefaniebutland. But I'm also hoping to present on the Rstudio conference

@llrs I just saw [your tweet about your RStudio talk](https://twitter.com/Lluis_Rev/status/1289301542659477504)! This is wonderful. I'll mark my calendar to contact you in late September about a post.

Enjoy your holidays!
