### 2020-08-05 1b82e04
Update documentation
More links and correct the reference page for pkgdown
