### 2020-08-05 54518ad
Update installation instructions
The most important link!!
