+++ README.Rmd
@@ -57,7 +57,7 @@ if (!require("BiocManager")) {
 You can install the latest version of BaseSet from [Github](https://github.com/ropensci/BaseSet) with:
 
 ```{r eval=FALSE}
-BiocManager::install("llrs/BaseSet", 
+BiocManager::install("ropensci/BaseSet", 
                      dependencies = TRUE, build_vignettes = TRUE, force = TRUE)
 ```
 
