### 2020-08-10 #359 @annakrystalli
Hey @llrs, everything looks good to me.  A couple of things to note:
- A couple of badges are showing as `unknown` (even though it looks like you've correctly set the new URLs). Might want to trigger a build to reanimate them.
- The CMD-CHECK GitHub Action & Travis seem to be failing. You might want to look into that.

Otherwise, I'm happy to conclude the review and close the issue. Now it's time for my holiday too! 😎🏝🍹
